% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-spec.R
\name{add_forest}
\alias{add_forest}
\title{Add a forest presentation of effect estimates}
\usage{
add_forest(x, axis = list(), width = 120, invert = FALSE)
}
\arguments{
\item{x}{A \verb{<mdl_gt>} specification.}

\item{axis}{Named options: \code{limits}, \code{breaks}, \code{intercept}, \code{log}, \code{title},
\code{left}, and \code{right}.}

\item{width}{Cell width in pixels.}

\item{invert}{Draw reciprocal effects.}
}
\description{
Add a forest presentation of effect estimates
}
