% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{addLibraryInformation}
\alias{addLibraryInformation}
\title{Add library metrics into the qseaSet sample table}
\usage{
addLibraryInformation(qseaSet)
}
\arguments{
\item{qseaSet}{A \code{qseaSet} object.}
}
\value{
The same \code{qseaSet}, with an augmented \code{sampleTable}.
}
\description{
Convenience helper to append columns from \code{qseaSet@libraries$file_name}
(and, when available, from \code{qseaSet@libraries$input_file} prefixed with
\code{"input_"}) to the \code{sampleTable}. Existing columns are not duplicated.
}
\examples{
data(exampleTumourNormal, package = "mesa")
qs <- addLibraryInformation(exampleTumourNormal)
head(qsea::getSampleTable(qs))

}
