% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQset.R
\name{addQseaNormalisationSteps}
\alias{addQseaNormalisationSteps}
\title{This function takes a qseaSet and adds several normalisation steps}
\usage{
addQseaNormalisationSteps(
  qseaSet,
  enrichmentMethod = "blind1-15",
  maxPatternDensity = 0.05,
  noTMM = TRUE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{enrichmentMethod}{What method to use to calculate the enrichment step}

\item{maxPatternDensity}{Maximum CG density in a window to consider it for the background calculation}

\item{noTMM}{Whether to normalise}
}
\value{
A qseaSet object with the samples merged together.
}
\description{
This function takes a qseaSet and adds several normalisation steps
}
