% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{arrange.qseaSet}
\alias{arrange.qseaSet}
\title{Arrange (reorder) samples in a qseaSet via dplyr syntax}
\usage{
\method{arrange}{qseaSet}(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A \code{qseaSet}.}

\item{...}{Variables to arrange by (passed to \code{\link[dplyr:arrange]{dplyr::arrange()}}).}

\item{.by_group}{Ignored; grouping is not used for \code{qseaSet}.}
}
\value{
The input \code{qseaSet} object with sample order updated.
}
\description{
Extends \code{\link[dplyr:arrange]{dplyr::arrange()}} to reorder \strong{samples} of a \code{qseaSet} based on columns
in its \code{sampleTable}.
}
\examples{
\donttest{
if (system.file("data","exampleTumourNormal.rda",package="mesa") != "") {
  data(exampleTumourNormal, package="mesa")
  exampleTumourNormal \%>\% arrange(age)
}
}
}
\seealso{
\code{\link[qsea:qseaSet-class]{qsea::getSampleTable()}}, \code{\link[=subsetQset]{subsetQset()}}
}
