% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{arrangeQset}
\alias{arrangeQset}
\title{This function takes a qseaSet and sorts the samples inside it using dplyr::arrange, using a column(s) of the sample table. Can also use desc() to reverse sort, as in dplyr::arrange.}
\usage{
arrangeQset(qseaSet, ..., .by_group = FALSE)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{...}{Other arguments to pass to dplyr::arrange}

\item{.by_group}{Not implemented as qsea requires a data frame not a tibble for the sampleTable.}
}
\value{
A qseaSet object with the samples sorted according to the column(s) selected.
}
\description{
This function takes a qseaSet and sorts the samples inside it using dplyr::arrange, using a column(s) of the sample table. Can also use desc() to reverse sort, as in dplyr::arrange.
}
