% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSgenomeDistributions.R
\docType{data}
\name{BSgenome.Hsapiens.UCSC.hg19.CpG.distribution}
\alias{BSgenome.Hsapiens.UCSC.hg19.CpG.distribution}
\alias{BSgenome.Hsapiens.NCBI.GRCh38.CpG.distribution}
\alias{BSgenome.Mmusculus.UCSC.mm10.CpG.distribution}
\title{Pre-calculated CpG distribution values for BSgenomes}
\format{
A data frame with 1 row and 2 columns

A data frame with 1 row and 2 columns

A data frame with 1 row and 2 columns
}
\usage{
BSgenome.Hsapiens.UCSC.hg19.CpG.distribution

BSgenome.Hsapiens.NCBI.GRCh38.CpG.distribution

BSgenome.Mmusculus.UCSC.mm10.CpG.distribution
}
\description{
A data frame containing cached relH and GoGe values for different BSgenomes, generated with calculateGenomicCGDistribution
}
\keyword{datasets}
