% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BSgenome.Hsapiens.NCBI.GRCh38.CpG.distribution.R
\docType{data}
\name{BSgenome.Hsapiens.NCBI.GRCh38.CpG.distribution}
\alias{BSgenome.Hsapiens.NCBI.GRCh38.CpG.distribution}
\title{Pre-calculated CpG distribution values for GRCh38}
\format{
A data frame with 1 rows and 2 variables:
}
\usage{
BSgenome.Hsapiens.NCBI.GRCh38.CpG.distribution
}
\description{
A data frame containing relH and GoGe values for GRCh38, generated with calculateGenomicCGDistribution
}
\keyword{datasets}
