% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{calculateArrayBetas}
\alias{calculateArrayBetas}
\title{This function calculates the beta values for the windows covering the hg38 array probes.}
\usage{
calculateArrayBetas(qseaSet)
}
\arguments{
\item{qseaSet}{qseaSet object to calculate the probe beta values for}
}
\description{
This function calculates the beta values for the windows covering the hg38 array probes.
}
