% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEnrichment.R
\name{calculateCpGEnrichment}
\alias{calculateCpGEnrichment}
\title{This function takes a path to a bam file and calculates the enrichment scores.}
\usage{
calculateCpGEnrichment(
  file = NULL,
  BSgenome = "BSgenome.Hsapiens.NCBI.GRCh38",
  exportPath = NULL,
  extend = 0,
  shift = 0,
  uniq = 0,
  chr.select = NULL,
  paired = TRUE
)
}
\arguments{
\item{file}{The path to the bam file}

\item{BSgenome}{Which BSGenome to use (GRCh38 is cached for speed)}

\item{exportPath}{Location to save a plot and data object into.}

\item{extend, shift, uniq, chr.select, paired}{Options to feed into medips::getGRange or medips::getPairedGRange}
}
\value{
A data frame containing the relH, GoGe and number of reads values for the samples
}
\description{
This function takes a path to a bam file and calculates the enrichment scores.
}
