% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEnrichment.R
\name{calculateCpGEnrichmentGRanges}
\alias{calculateCpGEnrichmentGRanges}
\title{This function takes a GRanges object and calculates the enrichment scores.}
\usage{
calculateCpGEnrichmentGRanges(
  readGRanges = NULL,
  BSgenome = NULL,
  chr.select = NULL
)
}
\arguments{
\item{readGRanges}{A GRanges object containing the span of each fragment of DNA}

\item{BSgenome}{Which BSgenome to use (GRCh38 is cached for speed)}

\item{chr.select}{Which chromosomes to use in global calculation}
}
\value{
A data frame containing the relH, GoGe and number of reads values for the samples
}
\description{
This function takes a GRanges object and calculates the enrichment scores.
}
