% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{colnames,qseaSet-method}
\alias{colnames,qseaSet-method}
\title{Column names of a qseaSet sample table}
\usage{
\S4method{colnames}{qseaSet}(x)
}
\arguments{
\item{x}{A \code{qseaSet}.}
}
\value{
Character vector of column names.
}
\description{
S4 method for BiocGenerics::colnames() that returns the column names of a
qseaSet's sample metadata table (qsea::getSampleTable(x)).
}
\examples{
\donttest{
if (system.file("data","exampleTumourNormal.rda",package="mesa") != "") {
  data(exampleTumourNormal, package="mesa")
  colnames(exampleTumourNormal)
}
}
}
\seealso{
qsea::getSampleTable, qsea::getSampleNames
}
