% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{colnames.qseaSet}
\alias{colnames.qseaSet}
\title{This function extends the base function length to return the number of samples in the qseaSet}
\usage{
\method{colnames}{qseaSet}(qseaSet)
}
\arguments{
\item{qseaSet}{A qseaSet to return the number of samples in}
}
\value{
The number of samples in the qseaSet
}
\description{
This function extends the base function length to return the number of samples in the qseaSet
}
