% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{countWindowsOverCutoff}
\alias{countWindowsOverCutoff}
\title{This function takes a qseaSet and sums over nrpm within a GRanges object}
\usage{
countWindowsOverCutoff(
  qseaSet,
  GRanges,
  samples = NULL,
  cutoff = 0,
  normMethod = "nrpm"
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{GRanges}{A GRanges object to sum over}

\item{samples}{What samples to use}

\item{cutoff}{What cutoff to use to count the windows above}

\item{normMethod}{What normalisation method to use}
}
\value{
A table
}
\description{
This function takes a qseaSet and sums over nrpm within a GRanges object
}
