% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropPooledControl}
\alias{dropPooledControl}
\title{This function drops the PooledControl sample from the qseaSet, if it is present.}
\usage{
dropPooledControl(qseaSet)
}
\arguments{
\item{qseaSet}{A qseaSet object to drop the PooledControl sample from}
}
\description{
This function drops the PooledControl sample from the qseaSet, if it is present.
}
