% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleMouse.R
\docType{data}
\name{exampleMouse}
\alias{exampleMouse}
\title{A small example qseaSet with 5 paired tumour/normal samples.}
\format{
A qseaSet object with 153 regions and 4 samples:
}
\usage{
exampleMouse
}
\description{
A qseaSet object containing a small region of chromosome 5 for a mouse cell
line, 2 replicates of 2 conditions (processed within CRUK-MI CBC).
}
\examples{
data("exampleMouse", package = "mesa")
exampleMouse
qsea::getSampleNames(exampleMouse)
}
\keyword{datasets}
