% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FantomRegions.R
\docType{data}
\name{FantomRegions}
\alias{FantomRegions}
\title{FANTOM5 regions for GRCh38}
\format{
A data frame with 63285 rows and 3 variables:
}
\usage{
FantomRegions
}
\description{
A data frame containing FANTOM5 enhancer regions, hg38
}
\keyword{datasets}
