% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{filterByOverlaps}
\alias{filterByOverlaps}
\title{Filter a qseaSet by overlaps with a GRanges object or table that can be coerced to one}
\usage{
filterByOverlaps(qseaSet, windowsToKeep)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{windowsToKeep}{A set of windows to keep, either a GRanges object, a dataframe with seqnames, start and end.}
}
\value{
A qseaSet object, with only a subset of the windows.
}
\description{
This function takes a qseaSet and keeps only a set of regions from it. Either specified as a GRanges object or an index vector of which regions.
}
