% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{filterRegions}
\alias{filterRegions}
\title{This function takes a qseaSet object and filters the regions inside it by a call to dplyr::filter.}
\usage{
filterRegions(qseaSet, ...)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{...}{Additional arguments to be used to filter the regions ONLY, as if they were a data frame.}
}
\value{
A qseaSet object, with the regions filtered appropriately.
}
\description{
This function takes a qseaSet object and filters the regions inside it by a call to dplyr::filter.
}
