% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Sample/group annotation data.frame}
\usage{
getAnnotation(qseaSet, useGroupMeans = FALSE, sampleAnnotation = NULL)
}
\arguments{
\item{qseaSet}{qsea::qseaSet.}

\item{useGroupMeans}{logical build annotations at group level.}

\item{sampleAnnotation}{tidyselect-style columns to include (unquoted).}
}
\value{
data.frame of annotations; rows named by \code{sample_name} (or \code{group}).
}
\description{
Select sample-table columns to use as annotations, either per-sample or
per-group (with consistency checks across replicates). Note that sampleAnnotation
must be enclosed within double curly brackets when used.
}
\examples{
data(exampleTumourNormal, package = "mesa")
qs <- exampleTumourNormal
ann <- getAnnotation(qs, useGroupMeans = FALSE, sampleAnnotation = group)
head(ann)

}
\seealso{
\code{\link[=makeHeatmapAnnotations]{makeHeatmapAnnotations()}}, \code{\link[=plotRegionsHeatmap]{plotRegionsHeatmap()}}, \code{\link[=plotCorrelationMatrix]{plotCorrelationMatrix()}}

Other heatmap-annotation: 
\code{\link{getAnnotationDataFrame}()},
\code{\link{getAnnotationDataFrameIndividual}()}
}
\concept{heatmap-annotation}
