% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getBetaMeans}
\alias{getBetaMeans}
\title{This function takes a qseaSet and calculates the mean of the nrpm and beta values over the windows given by GRanges.}
\usage{
getBetaMeans(qseaSet, GRanges = NULL, naMethod = "impute", minEnrichment = 3)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{GRanges}{A GRanges object with which windows to take the average of. If missing, all the regions in the qseaSet.}

\item{naMethod}{What method to use to deal with NA values for the beta values}

\item{minEnrichment}{The minimum number of reads for a window to be fully methylated, else replace with NA}
}
\value{
A table of data, one row per window
}
\description{
This function takes a qseaSet and calculates the mean of the nrpm and beta values over the windows given by GRanges.
}
