% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{getMesaAnnoDb}
\alias{getMesaAnnoDb}
\title{Get annotation DB for current or specified genome}
\usage{
getMesaAnnoDb(genome = NULL)
}
\arguments{
\item{genome}{Genome build, defaults to current setting}
}
\value{
Character string of annotation database name (e.g., "org.Hs.eg.db",
"org.Mm.eg.db")
}
\description{
Get annotation DB for current or specified genome
}
\examples{
getMesaAnnoDb("hg38")
getMesaAnnoDb("mm10")
}
