% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{getMesaTxDb}
\alias{getMesaTxDb}
\title{Get TxDb for current or specified genome}
\usage{
getMesaTxDb(genome = NULL)
}
\arguments{
\item{genome}{Genome build, defaults to current setting}
}
\value{
A TxDb object for the specified genome
}
\description{
Get TxDb for current or specified genome
}
\examples{
if (requireNamespace("TxDb.Hsapiens.UCSC.hg38.knownGene", quietly = TRUE)) {
    getMesaTxDb("hg38")
}
}
