% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getNormalisedReadSum}
\alias{getNormalisedReadSum}
\title{This function takes a qseaSet and sums over nrpm within a GRanges object}
\usage{
getNormalisedReadSum(qseaSet, GRanges, samples = NULL, subtractLevel = 0)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{GRanges}{A GRanges object to sum over}

\item{samples}{What samples to use}

\item{subtractLevel}{A value to subtract from the nrpm for each window before summing over them. Maybe useful to remove background noise.}
}
\value{
A table
}
\description{
This function takes a qseaSet and sums over nrpm within a GRanges object
}
