% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getPCAwithBatch}
\alias{getPCAwithBatch}
\title{This function edits the getPCA function from qsea to allow for batch effect correction}
\usage{
getPCAwithBatch(
  qs,
  chr = qsea::getChrNames(qs),
  minRowSum = 20,
  minEnrichment = 3,
  keep,
  normMethod = "beta",
  topVar = 1000,
  samples = qsea::getSampleNames(qs),
  batchVariable = NULL
)
}
\arguments{
\item{qs}{A qseaSet object}

\item{chr}{Which chromosomes to use}

\item{minRowSum}{Minimum number of reads in a window across all the samples to keep it}

\item{minEnrichment}{Minimal number of expected reads for a fully methylated window (for transformation to absolute methylation level)}

\item{keep}{a vector of row indices to keep (optional)}

\item{normMethod}{What normalisation method to use}

\item{topVar}{Number of most variable windows to keep}

\item{samples}{Which samples to use}

\item{batchVariable}{A variable in the sampleTable to batch correct for, using limma::removeBatchEffect}
}
\description{
This function edits the getPCA function from qsea to allow for batch effect correction
}
