% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{getUMAP}
\alias{getUMAP}
\title{This function is a modified version of the \code{\link[qsea:getPCA]{qsea::getPCA()}} function}
\usage{
getUMAP(
  qseaSet,
  dataTable = NULL,
  regionsToOverlap = NULL,
  normMethod = "beta",
  minEnrichment = 3,
  useGroupMeans = FALSE,
  minDensity = 0,
  topVarNum = 1000,
  topVarSamples = NULL,
  nPC = 2,
  minDist = 0.1,
  nNeighbours = 5,
  returnDataTable = FALSE,
  ...
)
}
\arguments{
\item{qseaSet}{A qseaSet object.}

\item{dataTable}{A data frame of normalised values for a set of windows (rows) and samples (columns), e.g. from \code{\link[=getDataTable]{getDataTable()}}. It must have seqnames, start and end columns. Can also be a \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}} object with normalised values in the metadata columns.}

\item{regionsToOverlap}{Optional. Only windows in \code{qseaSet} overlapping \code{regionsToOverlap} will be considered. A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}} object or a data frame which can be coerced into a \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}} object.}

\item{normMethod}{What normalisation method to use. Typically a character giving name of predefined normalisation method (e.g. "beta" or "nrpm"). See \code{\link[qsea:normMethod]{qsea::normMethod()}}.}

\item{minEnrichment}{Minimum number of reads for beta values to not give NA. Passed to \code{\link[=getDataTable]{getDataTable()}}.}

\item{useGroupMeans}{Whether to average samples over the group column (i.e. combine replicates)}

\item{minDensity}{A minimum CpG density level to filter out windows with values lower than.}

\item{topVarNum}{Number of most variable windows to keep. Defaults to 1000. If value is NA, NULL, Inf or is at least as big as the available number of windows, then all available windows are used for the PCA. Can also be a vector, in which case PCA is performed for each component.}

\item{topVarSamples}{Samples to use to determine variability. Either NULL or NA (use all samples; default), a character vector of sample names or a regular expression to match sample names on. Can also be a list, in which case PCA is performed for each list component. If \code{length(topVarNum) > 1}, list should be same length as \code{topVarNum} and each component will be matched with corresponding component of \code{topVarNum}.}

\item{nPC}{Number of components to be calculated. Default is 2.}

\item{minDist}{The effective minimum distance between embedded points. Smaller values will result in a more clustered/clumped embedding where nearby points on the manifold are drawn closer together, while larger values will result on a more even dispersal of points. The value should be set relative to the spread value, which determines the scale at which embedded points will be spread out.}

\item{nNeighbours}{The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation. Larger values result in more global views of the manifold, while smaller values result in more local data being preserved. In general values should be in the range 2 to 100.}

\item{returnDataTable}{A logical value indicating if the table of normalised values (generated by \code{\link[=getDataTable]{getDataTable()}} or by processing the input \code{dataTable}) should be returned as part of the output. See below for details.}

\item{...}{Other arguments to pass to uwot::umap.}
}
\value{
A list with the following components:
\item{pcas}{A named list of objects of class prcomp. List names correspond to entries in the \code{params$topVar} component of the output.}
\item{samples}{Names of the samples analysed.}
\item{windows}{A list of names of the windows analysed. List names correspond to entries in the \code{params$topVar} component of the output.}
\item{dataTable}{If \code{returnDataTable = TRUE}, the table of normalised values. The table will contain windows after filtering based on \code{regionsToOverlap} and \code{minDensity}, but prior to filtering based on \code{topVarNum}. Windows with missing values will also have been removed. The table will also include column(s) of window standard deviations, where applicable, with column name(s) corresponding to entries in the \code{params$topVar} component of the output.}
\item{params}{A list of the parameters used.}
\item{windowFiltering}{A list of the initial number of windows and the number filtered out at each filtering step (not including filtering based on \code{topVarNum}).}
}
\description{
This function is a modified version of the \code{\link[qsea:getPCA]{qsea::getPCA()}} function
}
