% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWindows}
\alias{getWindows}
\title{Extract the regions (windows) used in a qseaSet}
\usage{
getWindows(qseaSet)
}
\arguments{
\item{qseaSet}{A \code{qseaSet}.}
}
\value{
A \code{GRanges} of windows.
}
\description{
Convenience wrapper for \code{qsea::getRegions()}, returning the genomic windows.
}
\examples{
\donttest{
if (system.file("data","exampleTumourNormal.rda",package="mesa") != "") {
  data(exampleTumourNormal, package="mesa")
  getWindows(exampleTumourNormal)
}
}

}
\seealso{
\code{\link[qsea:qseaSet-class]{qsea::getRegions()}}, \code{\link[=getWindowNames]{getWindowNames()}}
}
