% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRCh38.99.gtf.R
\docType{data}
\name{GRCh38.99.gtf}
\alias{GRCh38.99.gtf}
\title{GTF file for GRCh38.}
\format{
A GRanges object with 1745514 ranges and 3 metadata columns:
\describe{
\item{type}{Type of feature (e.g. gene or exon)}
\item{gene_id}{ENSEMBL gene id}
\item{gene_name}{Common gene name}
...
}
}
\usage{
GRCh38.99.gtf
}
\description{
A GRanges object containing (some elements of) the v99 gtf of GRCh38
}
\keyword{datasets}
