% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg19ToHg38.over.chain.R
\docType{data}
\name{hg19ToHg38.over.chain}
\alias{hg19ToHg38.over.chain}
\title{hg19tohg38 liftover chain}
\format{
A chain object to liftOver hg19 to hg38
}
\usage{
hg19ToHg38.over.chain
}
\description{
A chain object to liftOver hg19 to hg38
}
\keyword{datasets}
