% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38_450kArrayGR.R
\docType{data}
\name{hg38_450kArrayGR}
\alias{hg38_450kArrayGR}
\title{Infinium 450k probe locations for hg38.}
\format{
A GRanges object with 485541 ranges and 2 metadata columns:
\describe{
\item{ID}{Array probe ID}
\item{MASK_general}{A field suggesting whether the probe could be unreliable (if TRUE) due to some kind of technical issue}
}
}
\usage{
hg38_450kArrayGR
}
\description{
A GRanges object containing the Illumina Infinium 450k array probe locations for hg38
}
\keyword{datasets}
