% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38CpGIslands.R
\docType{data}
\name{hg38CpGIslands}
\alias{hg38CpGIslands}
\title{CpG islands for hg38.}
\format{
A GRanges object with 31144 ranges and 7 metadata columns:
\describe{
\item{length}{Number of bases in each island}
\item{cpgNum}{Number of CGs in each island}
\item{gcNum}{Number of Cs or Gs in the island}
\item{perCpg}{Percent CGs in the island}
\item{perGc}{Percent of Cs or Gs in the island}
\item{obsExp}{Unclear what this column means}
...
}
}
\usage{
hg38CpGIslands
}
\description{
A GRanges object containing CpG islands in hg38 coordinates. Generated from http://hgdownload.cse.ucsc.edu/goldenpath/hg38/database/cpgIslandExt.txt.gz
}
\keyword{datasets}
