% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{keepWindowsOverCutoff}
\alias{keepWindowsOverCutoff}
\title{This function takes a qseaSet and keeps the windows which have expression above a cutoff in a defined set of samples}
\usage{
keepWindowsOverCutoff(qseaSet, samplesToFilterOn, minValue, normMethod)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{samplesToFilterOn}{A set of sample names to keep, or a string to match in the sample names}

\item{minValue}{The minimum value to require in any of the samples}

\item{normMethod}{The type of normalisation method to use (nrpm, beta, count)}
}
\value{
A qseaSet object, with only a subset of the windows.
}
\description{
This function takes a qseaSet and keeps the windows which have expression above a cutoff in a defined set of samples
}
