% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{leftJoinQset}
\alias{leftJoinQset}
\title{This function takes a qseaSet and appends data to its sampleTable based on a call to dplyr::left_join}
\usage{
leftJoinQset(qseaSet, newData, ...)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{newData}{The new data frame to join the sampleTable on.}

\item{...}{Other arguments to pass to dplyr::left_join. Can include "by" to specify which columns to join on}
}
\value{
A qseaSet object with the sampleTable appended to as specified.
}
\description{
This function takes a qseaSet and appends data to its sampleTable based on a call to dplyr::left_join
}
