% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mesaDimRed}
\alias{mesaDimRed}
\title{Constructor function of a mesaDimRed object.}
\usage{
mesaDimRed(res, sampleTable, samples, params, dataTable = data.frame())
}
\arguments{
\item{res}{The results, a list containing individual PCA/UMAP objects.}

\item{sampleTable}{A sampleTable from the qseaSet.}

\item{samples}{A character vector of the samples used.}

\item{params}{The parameters used during construction.}

\item{dataTable}{The data used to construct the PCA/UMAP object (optional).}
}
\value{
A \code{mesaDimRed} object.
}
\description{
Constructor function of a mesaDimRed object.
}
