% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mesaPCA}
\alias{mesaPCA}
\title{Constructor function of a mesaPCA object.}
\usage{
mesaPCA(prcomp, windows)
}
\arguments{
\item{prcomp}{A principal component object output from \code{\link[stats:prcomp]{stats::prcomp()}}}

\item{windows}{A character vector of the windows used.}
}
\value{
A \code{mesaPCA} object.
}
\description{
Constructor function of a mesaPCA object.
}
