% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{mesaSSROptions}
\alias{mesaSSROptions}
\title{Default SSR options}
\format{
An object of class \code{list} of length 3.
}
\usage{
mesaSSROptions
}
\description{
Default options for server-side rendering. These are the default values used
in \link{mesa}. To modify an option, supply a named list with the option you would
like to change. You only need to supply the values to modify.
}
\details{
\itemize{
\item \code{useSSR} Should server-side rendering be used? If \code{FALSE}, the remaining
options are ignored.
\item \code{usePagination} Should server-side rendering be done via pagination?
\item \code{useInfiniteScroll} Should server-side rendering be done via infinite scroll?
}

If \code{useSSR} is \code{FALSE}, the remaining arguments are ignored. If both
\code{usePagination} and \code{useInfiniteScroll} are \code{TRUE}, infinite scroll will be
used.
}
\keyword{datasets}
