% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mesaUMAP}
\alias{mesaUMAP}
\title{Constructor function of a mesaUMAP object.}
\usage{
mesaUMAP(points, windows)
}
\arguments{
\item{points}{Data frame containing one row for each sample, with the positions in UMAP space.}

\item{windows}{A character vector of the windows used.}
}
\value{
A \code{mesaUMAP} object.
}
\description{
Constructor function of a mesaUMAP object.
}
