% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{mixArrayWithQset}
\alias{mixArrayWithQset}
\title{This function takes a qseaSet and makes a new sample by mixing two samples}
\usage{
mixArrayWithQset(
  qseaSet,
  arrayReadTable,
  arraySample,
  qseaSample,
  nReadsTotal,
  proportion,
  newName = NULL,
  groupName = NULL,
  onlyNew = FALSE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{arrayReadTable}{Data frame with the array probe beta values}

\item{arraySample}{Name of the array sample to mix}

\item{qseaSample}{Name of the qseaSet sample to mix with}

\item{nReadsTotal}{Number of reads in total to have after mixing}

\item{proportion}{The proportion to take from sample1, the rest will come from sample2}

\item{newName}{A name to give the new sample}

\item{groupName}{A name to use in the group column in the sampleTable}

\item{onlyNew}{Whether to only return the new sample.}
}
\value{
A qseaSet object with an extra
}
\description{
This function takes a qseaSet and makes a new sample by mixing two samples
}
