% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{mutateQset}
\alias{mutateQset}
\title{This function takes a qseaSet and mutates its sampleTable based on a call to dplyr::mutate.}
\usage{
mutateQset(qseaSet, ...)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{...}{Other arguments to pass to dplyr::mutate}
}
\value{
A qseaSet object with the sampleTable mutated as specified.
}
\description{
This function takes a qseaSet and mutates its sampleTable based on a call to dplyr::mutate.
}
