% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
The result of applying \code{rhs} to \code{lhs}.
}
\description{
Re-export of the magrittr pipe. See
\code{\link[magrittr:pipe]{\%>\%}} for full details.
It forwards the left-hand side (LHS) as the first argument to the
right-hand side (RHS) call.
}
\examples{
# Basic usage
1:5 \%>\% sum()

# Chain multiple transformations
1:10 \%>\%
  mean() \%>\%
  round(1)
}
\keyword{internal}
