% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{Sample correlation heatmap}
\usage{
plotCorrelationMatrix(
  qseaSet,
  regionsToOverlap = NULL,
  useGroupMeans = FALSE,
  sampleAnnotation = NULL,
  normMethod = "nrpm",
  minEnrichment = 3,
  annotationColors = NA,
  minDensity = 0,
  ...
)
}
\arguments{
\item{qseaSet}{qsea::qseaSet.}

\item{regionsToOverlap}{optional GRanges to restrict windows.}

\item{useGroupMeans}{logical average replicates by \code{group}.}

\item{sampleAnnotation}{tidyselect-style sample-table columns to annotate.}

\item{normMethod}{character(1) \code{"nrpm"} or \code{"beta"}.}

\item{minEnrichment}{integer minimum reads for non-NA beta.}

\item{annotationColors}{optional list of colours for annotations (pheatmap).}

\item{minDensity}{numeric minimum CpG density to keep windows.}

\item{...}{passed to \code{pheatmap::pheatmap()}.}
}
\value{
A \code{pheatmap} object.
}
\description{
Compute and plot the correlation matrix across samples (or group means)
using the selected normalisation method and optional region filters.
}
\examples{
\donttest{
data(exampleTumourNormal, package = "mesa")
qs <- exampleTumourNormal
plotCorrelationMatrix(qs, normMethod = "nrpm", sampleAnnotation = group)
}

}
\seealso{
\code{\link[=plotRegionsHeatmap]{plotRegionsHeatmap()}}, \code{\link[=getDataTable]{getDataTable()}}

Other heatmaps: 
\code{\link{plotGeneHeatmap}()},
\code{\link{plotRegionsHeatmap}()}
}
\concept{heatmaps}
