% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotQseaPCA}
\alias{plotQseaPCA}
\title{This function takes a qseaSet and plots a PCA}
\usage{
plotQseaPCA(
  qseaSet,
  signatureGR = NULL,
  plotComponents = c(1, 2),
  plotColour = "type",
  plotShape = "experiment",
  batchVariable = NULL,
  showNames = FALSE,
  normMethod = "beta",
  minRowSum = 20,
  minEnrichment = 3,
  topVar = NULL,
  plotTitle = "",
  returnDataOnly = FALSE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{signatureGR}{A GRanges (or data frame coercible to one) with a subset of windows to calculate the PCA on.}

\item{plotComponents}{Which of the PCA components to plot default c(1,2)}

\item{plotColour}{Column of the sampleTable to use to set the colour of the points}

\item{plotShape}{Column of the sampleTable to use to set the shape of the points}

\item{batchVariable}{A variable to use for batch correction using limma. Can be overaggressive and remove all variation. Be careful with this!}

\item{showNames}{Whether to show the names of the points}

\item{normMethod}{Normalisation method to use, nrpm or beta.}

\item{minRowSum}{A minimum number of reads to be included in the PCA}

\item{minEnrichment}{Minimal number of expected reads for a fully methylated window (for transformation to absolute methylation level)}

\item{topVar}{Take only the n most variable windows in the PCA. If NULL, all valid windows.}

\item{plotTitle}{Title for the plot}

\item{returnDataOnly}{Return just the data frame behind the PCA}
}
\value{
A plot of the PCA
}
\description{
This function takes a qseaSet and plots a PCA
}
