% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_and_capture}
\alias{print_and_capture}
\title{This function captures a printed form of the object x, for use in error messages
Taken from https://stackoverflow.com/a/26083626 by Richie Cotton}
\usage{
print_and_capture(x)
}
\arguments{
\item{x}{An object to capture}
}
\description{
This function captures a printed form of the object x, for use in error messages
Taken from https://stackoverflow.com/a/26083626 by Richie Cotton
}
