% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{pullQset}
\alias{pullQset}
\title{This function takes a qseaSet and pulls a column from its sampleTable based on a call to dplyr::pull}
\usage{
pullQset(qseaSet, ...)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{...}{Other arguments to pass to dplyr::pull}
}
\value{
A column of the sampleTable as a vector
}
\description{
This function takes a qseaSet and pulls a column from its sampleTable based on a call to dplyr::pull
}
