% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseDMRs.R
\name{qseaWriteDMRsExcel}
\alias{qseaWriteDMRsExcel}
\title{Write the qsea result to an Excel file}
\usage{
qseaWriteDMRsExcel(dataTable, path, fdrThres = 0.05)
}
\arguments{
\item{dataTable}{An (optionally annotated) data table with contrasts}

\item{path}{The path to write the excel file to.}

\item{fdrThres}{FDR rate threshold to apply to each contrast}
}
\value{
Returns the original table invisibly, so can be used in a pipe.
}
\description{
This writes an Excel file with the results, with one sheet per contrast.
}
