% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaCNV.R
\name{removeCNV}
\alias{removeCNV}
\title{Remove (zero out) CNV data from a qseaSet}
\usage{
removeCNV(qseaSet)
}
\arguments{
\item{qseaSet}{A \code{qseaSet}.}
}
\value{
A \code{qseaSet} with the CNV assay zeroed.
}
\description{
Sets all per-window CNV values to zero, keeping genomic coordinates intact.
Useful for resetting CNV prior to recalculation.
}
\examples{
# data(exampleTumourNormal, package = "mesa")
# qs <- exampleTumourNormal
# qs <- removeCNV(qs)
# qsea::getCNV(qs)

}
\seealso{
\code{\link{addHMMcopyCNV}}, \code{\link{plotCNVheatmap}}

Other CNV: 
\code{\link{addHMMcopyCNV}()},
\code{\link{plotCNVheatmap}()},
\code{\link{runHMMCopy}()}
}
\concept{CNV}
