% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{removeWindowsOverCutoff}
\alias{removeWindowsOverCutoff}
\title{This function takes a qseaSet and filters out windows which have expression in a defined set of samples}
\usage{
removeWindowsOverCutoff(
  qseaSet,
  samplesToFilterOut,
  maxValue,
  normMethod,
  .swap = FALSE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{samplesToFilterOut}{A set of sample names to filter out, or a string to match in the sample names}

\item{maxValue}{The maximum value to allow in any of the samples}

\item{normMethod}{The type of normalisation method to use (nrpm, beta, count)}

\item{.swap}{Whether to change the printed messages as we are actually going to be keeping these windows rather than filtering them.}
}
\value{
A qseaSet object, with only a subset of the windows.
}
\description{
This function takes a qseaSet and filters out windows which have expression in a defined set of samples
}
