% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{renameQsetNames}
\alias{renameQsetNames}
\title{Rename samples in a qseaSet by regex}
\usage{
renameQsetNames(qseaSet, pattern, replacement = "")
}
\arguments{
\item{qseaSet}{A \code{qseaSet} object.}

\item{pattern}{A regex \code{pattern} to match in sample names.}

\item{replacement}{A replacement string (default: \code{""}).}
}
\value{
A \code{qseaSet} with renamed samples.
}
\description{
Apply a regex replacement to sample names and update all relevant slots
consistently (counts, CNV, enrichment, zygosity, libraries, sampleTable).
}
\examples{
data(exampleTumourNormal, package = "mesa")
# Replace "T" with "Tumour" in sample names
qs2 <- renameQsetNames(exampleTumourNormal, pattern = "T$", replacement = "Tumour")
head(colnames(qs2@count_matrix))

}
