% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{renameSamples}
\alias{renameSamples}
\title{Rename samples using a column from the sample table}
\usage{
renameSamples(qseaSet, newNameColumn)
}
\arguments{
\item{qseaSet}{A \code{qseaSet} object.}

\item{newNameColumn}{Name (unquoted) of a column in the sample table that
contains the new sample names. Must be unique and contain no \code{NA}s.}
}
\value{
A \code{qseaSet} with sample names replaced by \code{newNameColumn}.
}
\description{
Replace sample names with values from a user‑supplied column in
\code{qseaSet@sampleTable}, updating all dependent slots consistently.
}
\examples{
data(exampleTumourNormal, package = "mesa")
qs <- exampleTumourNormal \%>\%
  dplyr::mutate(newName = paste0("Sample", seq_len(nrow(qsea::getSampleTable(exampleTumourNormal)))))
qs_renamed <- renameSamples(qs, newNameColumn = newName)
head(qsea::getSampleTable(qs_renamed)$sample_name)

}
