% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesa.R
\name{renderMesa}
\alias{renderMesa}
\title{Shiny render function}
\usage{
renderMesa(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a \link{mesa} table.}

\item{env}{The environment in which to evaluate \code{expr.}}

\item{quoted}{Is \code{expr} a quoted expression?}
}
\description{
Shiny render function
}
