% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{selectQset}
\alias{selectQset}
\title{Select columns of a qseaSet sample table}
\usage{
selectQset(qseaSet, ...)
}
\arguments{
\item{qseaSet}{A \code{qseaSet}.}

\item{...}{Selection helpers passed to \code{\link[dplyr:select]{dplyr::select()}}.}
}
\value{
The input \code{qseaSet} object with \code{sampleTable} updated.
}
\description{
Keeps/renames columns from the \code{sampleTable} of a \code{qseaSet}. If present,
\code{sample_name} and \code{group} are kept at the front of the table.
}
\examples{
\donttest{
if (system.file("data","exampleTumourNormal.rda",package="mesa") != "") {
  data(exampleTumourNormal, package="mesa")
  selectQset(exampleTumourNormal, sample_name, group)
}
}
}
\seealso{
\code{\link[=select.qseaSet]{select.qseaSet()}}, \code{\link[qsea:qseaSet-class]{qsea::getSampleTable()}}
}
