% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_long_checks}
\alias{skip_long_checks}
\title{This function skips long running tests if options(skip_long_checks = TRUE) has been set}
\usage{
skip_long_checks()
}
\description{
This function skips long running tests if options(skip_long_checks = TRUE) has been set
}
