% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{sortQset}
\alias{sortQset}
\title{This function takes a qseaSet and sorts the samples in it, so they are in alphabetical order (with numbers increasing via mixedsort)}
\usage{
sortQset(qseaSet, decreasing = FALSE)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{decreasing}{Whether to order the qseaSet in reverse alphabetical order}
}
\value{
A qseaSet object with the objects sorted
}
\description{
This function takes a qseaSet and sorts the samples in it, so they are in alphabetical order (with numbers increasing via mixedsort)
}
