% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{subsetQset}
\alias{subsetQset}
\title{Subset a qseaSet by samples}
\usage{
subsetQset(qseaSet, samplesToKeep = NULL, samplesToDrop = NULL)
}
\arguments{
\item{qseaSet}{A \code{qseaSet} object.}

\item{samplesToKeep}{Character vector of sample names to keep. Use \strong{either}
\code{samplesToKeep} \strong{or} \code{samplesToDrop}.}

\item{samplesToDrop}{Character vector of sample names to drop. Use \strong{either}
\code{samplesToKeep} \strong{or} \code{samplesToDrop}.}
}
\value{
A \code{qseaSet} containing only the selected samples.
}
\description{
Keep only a specified set of samples (or, alternatively, drop a set). All
internal matrices/slots are subset consistently.
}
\examples{
data(exampleTumourNormal, package = "mesa")
subsetQset(exampleTumourNormal, samplesToKeep = c("Colon1_T","Colon1_N"))

}
