% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-forest.R
\name{tbl_forest}
\alias{tbl_forest}
\alias{tbl_interaction_forest}
\title{Forest plot and table}
\usage{
tbl_interaction_forest(
  object,
  outcomes = formula(),
  exposures = formula(),
  interactions = formula(),
  level_labels = formula(),
  columns = list(beta ~ "Estimate", conf ~ "95\% CI", n ~ "No."),
  axis = list(scale ~ "continuous"),
  width = list(),
  forest = list(),
  exponentiate = FALSE,
  invert = FALSE,
  digits = 2,
  ...
)
}
\arguments{
\item{object}{A \verb{<mdl_tbl>} object with the required models. It must also
contain the original dataset used to create the models of interest. Please
see \code{\link[=attach_data]{attach_data()}} for details.}

\item{outcomes}{A \verb{<formula>} or list of formulas selecting the outcome
variables of interest. The \strong{LHS} is always the name of the variable that
will be selected. The \strong{RHS} is the potential label for the output table.}

\item{exposures}{A \verb{<formula>} or list of formulas of strata that should be
evaluated, with the \strong{LHS} referring to the strata and the \strong{RHS}
referring to its label.}

\item{interactions}{A \verb{<formula>} or list of formulas of the interaction
terms that should be evaluated, with the \strong{LHS} referring to the term and
the \strong{RHS} referring to its label. Currently only supports binary
interaction terms.}

\item{level_labels}{A \verb{<formula>} or list of formulas where each list-element
is a formula with the \strong{LHS} reflecting either the variable to re-label or
a specific level, and the \strong{RHS} reflecting what the new level should be
called (for display). If there are conflicting labels, the most recent will
be used.

For example, \verb{list(am ~ c("Manual", "Automatic")} would take, from the
\code{mtcars} dataset, the \code{am} variable, which consists of \code{c(0, 1)}, and
relabel them in the order described. They are sorted in ascendinng order
prior to re-labeling.

The alternative approach is to use the specific level itself and have it
re-labeled. \code{list(0 ~ "Absent")} would take all levels that are zero, and
change their value.}

\item{columns}{Columns that help to describe the individual terms of a model.
Generally describes the statistical estimates or properties of an  exposure
term. At least one column should likely be selected from this list. The
sequence listed will reflect the sequence shown in the table. This is given
as a \verb{<formula>} or list of formulas, with the \strong{LHS} is the name of an
acceptable variable, and the \strong{RHS} is the potential column label. The
current options are:
\itemize{
\item beta = point estimate value, such as odds ratio or hazard ratio
\item conf = inclusion of the confidence interval (presumed to be ~95\%-ile)
\item n = number of observations in each group or subset
\item p = p_value for model or interaction term
}

For example: \verb{list(beta ~ "Hazard", conf ~ "95\% CI" n ~ "No.")"}

Notably, the columns do not describe the terms, but the data contained in
the rows below. The \code{terms} and \code{level_labels} arguments serve as the
spanning labels for the related columns. Please see \code{\link[gt:tab_spanner]{gt::tab_spanner()}} for
more information on how spanners work.}

\item{axis}{Argument to help modify the forest plot itself. This is a
\verb{<formula>} or list of formulas of the following parameters. If they are
not named, the function will attempt to "guess" the optimal parameters. The
options are:
\itemize{
\item title = label or title for the column describing the forest plot
\item lim = x-axis limits
\item breaks = x-axis tick marks or break points that should be numbers
\item int = x-axis intercept
\item lab = label for the x-axis
\item scale = defaults to continuous, but may also use a log transformation as
well \code{c("continuous", "log")}
}

For example: \verb{list(title ~ "Decreasing Hazard", lab ~ "HR (95\% CI))}}

\item{width}{Describes the width of each column in a \verb{<formula>} or list of
formulas. The \strong{RHS} is a decimal reflecting the percent each column
should take of the entire table. The forest plot is usually given 30\% of
the width. The default options attempt to be sensible. Options, indicated
by the term on the -\emph{LHS}* of the formula, include:
\itemize{
\item n = Column describing number of observations
\item beta = Column of estimate and confidence intervals (usually combined)
\item forest = Column containing forest plots
}

For example: \code{list(n ~ .1, forest ~ 0.3)}}

\item{forest}{A \verb{<formula>} or list of formulas that can be used to help
customize the forest plot prior to generation of the table. The options
directly correspond to \code{ggplot2} aesthetic specifications that can modify
the visual aspects of the forest plot. The currently supported arguments:
\itemize{
\item size = Relative size of the marker for point estimate
\item shape = Shape of the marker for point estimate
\item fill = Fill of the marker for point estimate
\item linetype = Vertical line that serves as the x-intercept across the table
\item linewidth = Thickness of lines, for both vertical and horixontal axes
}}

\item{exponentiate}{A \verb{<logical>} value that determines if the point estimates
should be exponentiated. This is useful for odds ratios or hazard ratios. It
will apply to all estimate values, including confidence intervals, in a
table. It is defaulted to \code{FALSE}.}

\item{invert}{A \verb{<logical>} to determine if the odds or hazard ratio should
be shown as the reciprocal values. Instead of a decreasing hazard for every
unit increase, it describes an increasing hazard for every unit decrease.
Default is \code{FALSE}}

\item{digits}{The number of significant figures to present. If the numbers
are not scaled in a presentable fashion, can always adjust the table
subsequently.}
}
\description{
Forest plots are usually ways to describe contrasting data, such as between
strata, or to show interaction (if present). We can show the estimates of
each parameter along a dichotomous subgroup, or we can show the estimates of
a primary exposure along a multitude of subgroups. This function allows both
methods (and some spectrum in between) to demonstrate these.
}
