% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tbls}
\alias{tbls}
\title{Extensions of \code{{gt}} package for multiple models}
\arguments{
\item{object}{A \verb{<mdl_tbl>} object with the required models. It must also
contain the original dataset used to create the models of interest. Please
see \code{\link[=attach_data]{attach_data()}} for details.}

\item{data}{The original dataset used to create the models of interest. The
variables in the model must be contained within this dataset. This may also
be attached}

\item{outcomes}{A \verb{<formula>} or list of formulas selecting the outcome
variables of interest. The \strong{LHS} is always the name of the variable that
will be selected. The \strong{RHS} is the potential label for the output table.}

\item{exposures}{A \verb{<formula>} or list of formulas of strata that should be
evaluated, with the \strong{LHS} referring to the strata and the \strong{RHS}
referring to its label.}

\item{interactions}{A \verb{<formula>} or list of formulas of the interaction
terms that should be evaluated, with the \strong{LHS} referring to the term and
the \strong{RHS} referring to its label. Currently only supports binary
interaction terms.}

\item{strata}{A \verb{<formula>} or list of formulas of strata that should be
evaluated, with the \strong{LHS} referring to the strata and the \strong{RHS}
referring to its label.}

\item{level_labels}{A \verb{<formula>} or list of formulas where each list-element
is a formula with the \strong{LHS} reflecting either the variable to re-label or
a specific level, and the \strong{RHS} reflecting what the new level should be
called (for display). If there are conflicting labels, the most recent will
be used.

For example, \verb{list(am ~ c("Manual", "Automatic")} would take, from the
\code{mtcars} dataset, the \code{am} variable, which consists of \code{c(0, 1)}, and
relabel them in the order described. They are sorted in ascendinng order
prior to re-labeling.

The alternative approach is to use the specific level itself and have it
re-labeled. \code{list(0 ~ "Absent")} would take all levels that are zero, and
change their value.}

\item{terms}{A \verb{<formula>} or list of formulas selecting the model terms that
should be used. The \strong{LHS} is always the name of the variable that will be
selected. The \strong{RHS} is the potential label for the output table.}

\item{columns}{Columns that help to describe the individual terms of a model.
Generally describes the statistical estimates or properties of an  exposure
term. At least one column should likely be selected from this list. The
sequence listed will reflect the sequence shown in the table. This is given
as a \verb{<formula>} or list of formulas, with the \strong{LHS} is the name of an
acceptable variable, and the \strong{RHS} is the potential column label. The
current options are:
\itemize{
\item beta = point estimate value, such as odds ratio or hazard ratio
\item conf = inclusion of the confidence interval (presumed to be ~95\%-ile)
\item n = number of observations in each group or subset
\item p = p_value for model or interaction term
}

For example: \verb{list(beta ~ "Hazard", conf ~ "95\% CI" n ~ "No.")"}

Notably, the columns do not describe the terms, but the data contained in
the rows below. The \code{terms} and \code{level_labels} arguments serve as the
spanning labels for the related columns. Please see \code{\link[gt:tab_spanner]{gt::tab_spanner()}} for
more information on how spanners work.}

\item{accents}{A \verb{<formula>} or list of formulas that provide instructions
on how to emphasize or accentuate certain data cells within the table.
Similar to the \code{columns} argument, the \strong{LHS} refers to the data available
in the columns, such as the \emph{p} or \emph{beta} value, but is stated as a
criteria, e.g. \code{p < 0.05}. The \strong{RHS} is the instruction on what type of
accentuation to perform. This applies to \emph{ALL} data columns of the table,
if the criteria are met.

The \strong{LHS} options are:
\itemize{
\item beta = point estimate value
\item conf_low = lower bound of the confidence interval
\item conf_high = upper bound of the confidence interval
\item p = p-value for a model term
}

The current \strong{RHS} options are:
\itemize{
\item bold = apply bold labeling to text
}}

\item{suppress_column_labels}{A \verb{<logical>} value that determines if the
column labels should be suppressed and replaced by the term labels.

This defaults to \code{TRUE} when only a single column label is chosen. In this
case, the individual term level labels are used, which are obtained either
explicitly through the \strong{level_labels} argument or through the default
levels from the original data.

If multiple columns are chosen to be displayed, then the column labels by
default are retained (argument is \code{FALSE}). Occasionally, several columns
will be initially selected and subsequently \emph{hidden} by the user such that
only a single column (per term) remains. This is a use case to force the
column labels to be suppressed.}

\item{exponentiate}{A \verb{<logical>} value that determines if the point estimates
should be exponentiated. This is useful for odds ratios or hazard ratios. It
will apply to all estimate values, including confidence intervals, in a
table. It is defaulted to \code{FALSE}.}
}
\description{
The related functions in this package combine multiple models to help show
the relationship between variables.
}
\details{
The extensions that combine the modeling functions of the \code{{vlndr}} package
and table-making functions of the \code{{gt}} package require some conventions to
simplify the process. These conventions are listed and explained below.
}
\section{Formula inputs}{
When providing pluralized arguments, such as for \strong{outcomes}, \strong{exposures},
\strong{interactions}, and \strong{strata}, both the variable itself and its potential
label are of interest. These \emph{labeled formulas} are handled generally by the
function \code{\link[=labeled_formulas_to_named_list]{labeled_formulas_to_named_list()}} The input types are (in order of
preference/informativity):
\itemize{
\item \verb{<list>} of \verb{<formula>} objects: The \strong{LHS} is always the name of the variable that will be selected. The \strong{RHS} is the potential label for the output table.
\item \verb{<formula>} object
\item \verb{<character>} vector
}
}

