% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{testPlotGeneHeatmap}
\alias{testPlotGeneHeatmap}
\title{An internal wrapper around plotGeneHeatmap that catches biomart connection errors and skips the test if they occur}
\usage{
testPlotGeneHeatmap(...)
}
\arguments{
\item{...}{Arguments to pass to plotGeneHeatmap}
}
\description{
An internal wrapper around plotGeneHeatmap that catches biomart connection errors and skips the test if they occur
}
