% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseDMRs.R
\name{writeDMRsToExcel}
\alias{writeDMRsToExcel}
\title{Write DMR results to Excel}
\usage{
writeDMRsToExcel(dataTable, path, FDRthres = 0.05)
}
\arguments{
\item{dataTable}{A \code{data.frame} of DMR results, typically from \code{\link[=calculateDMRs]{calculateDMRs()}}.}

\item{path}{\code{character(1)} File path to write the Excel workbook.}

\item{FDRthres}{\code{numeric(1)} False discovery rate threshold for filtering.}
}
\value{
Invisibly returns the input \code{dataTable}, enabling use in a pipe.
}
\description{
Export DMR results to an Excel workbook with one sheet per contrast.
}
\examples{
\dontrun{
data(exampleTumourNormal, package = "mesa")
exampleTumourNormal \%>\%
  calculateDMRs(variable = "type", contrasts = "LUAD_vs_NormalLung") \%>\%
  writeDMRsToExcel(path = "dmr_results.xlsx")
  
exampleTumourNormal \%>\%
  calculateDMRs(variable = "tumour", contrasts = "all") \%>\%
  annotateWindows(TxDb = "TxDb.Hsapiens.UCSC.hg38.knownGene", annoDb = "org.Hs.eg.db") \%>\%
  writeDMRsToExcel(paste0(tempdir(),"/test.xlsx"))
}

}
\seealso{
\code{\link[=writeDMRsToBed]{writeDMRsToBed()}}

Other DMR-helpers: 
\code{\link{pivotDMRsLonger}()},
\code{\link{summariseDMRsByContrast}()},
\code{\link{summariseDMRsByGene}()},
\code{\link{writeDMRsToBed}()}
}
\concept{DMR-helpers}
