% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQset.R
\name{addBamCoveragePairedAndUnpaired}
\alias{addBamCoveragePairedAndUnpaired}
\title{This function adds coverage to the qseaSet, using both paired reads and the high-quality R1s.}
\usage{
addBamCoveragePairedAndUnpaired(
  qs,
  fragmentLength = NULL,
  minMapQual = 30,
  maxInsertSize = 1000,
  minInsertSize = 100,
  minReferenceLength = 30,
  parallel = FALSE,
  properPairsOnly = FALSE
)
}
\arguments{
\item{qs}{qseaSet object to add coverage to.}

\item{fragmentLength}{Length to extend unpaired reads to}

\item{minMapQual}{Minimum mapping quality to include a read (on either end if proper pair)}

\item{maxInsertSize}{Maximum insert size for a proper paired read}

\item{minInsertSize}{Minimum insert size for a proper paired read}

\item{minReferenceLength}{Minimum length on the reference coordinates for a non-paired read to be included}

\item{parallel}{Whether to use parallelisation}

\item{properPairsOnly}{Whether to only keep proper pairs, else high quality unpaired R1s are also kept.}
}
\value{
A qseaSet with the coverage added.
}
\description{
This function adds coverage to the qseaSet, using both paired reads and the high-quality R1s.
}
