% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaCNV.R
\name{addHMMcopyCNV}
\alias{addHMMcopyCNV}
\title{Calculate and add CNV based on HMMcopy, on individual samples}
\usage{
addHMMcopyCNV(
  qs,
  inputColumn = "input_file",
  windowSize = 1000000,
  fragmentLength = NULL,
  plotDir = NULL,
  parallel = getMesaParallel(),
  maxInsertSize = 1000,
  minInsertSize = 50,
  minReferenceLength = 30,
  minMapQual = 30,
  properPairsOnly = FALSE,
  hmmCopyGC = NULL,
  hmmCopyMap = NULL
)
}
\arguments{
\item{qs}{The qseaSet object}

\item{inputColumn}{Which column of the sample table contains the link to the bam file}

\item{windowSize}{What window size to use (only 50000, 500000 or 1000000 work at the moment)}

\item{fragmentLength}{Fragment length to extend reads to if not paired}

\item{plotDir}{A directory to export individual HMMcopy plots to.}

\item{parallel}{Whether to use parallelisation}

\item{maxInsertSize}{For paired reads, only keep them if they are below a maximum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.}

\item{minInsertSize}{For paired reads, only keep them if they are above a minimum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.}

\item{minReferenceLength}{A minimum distance on the genome to keep the read. bwa by default gives 19bp as minimum for a read, which is quite short.}

\item{minMapQual}{Minimum mapping quality to include a read (on either end if proper pair)}

\item{properPairsOnly}{Whether to only keep properly paired reads, or to keep high-quality (MAPQ 30+) unpaired R1s as well. Set to TRUE for size selection.}

\item{hmmCopyGC}{Object containing GC content per window in the genome, for hmmcopy}

\item{hmmCopyMap}{Object containing mapability content per window in the genome, for hmmcopy}
}
\value{
A qseaGLM object
}
\description{
This calculates and adds CNV to a qseaSet based on just using HMMcopy on each sample, with default parameters.
Only works for hg38/GRCh38!
}
