% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQC.R
\name{addHyperStableFraction}
\alias{addHyperStableFraction}
\title{This function calculates and summarises the proportion of hyperstable regions that are present in each sample, and adds to columns of the sampleTable}
\usage{
addHyperStableFraction(qseaSet)
}
\arguments{
\item{qseaSet}{A qseaSet object}
}
\description{
This function calculates and summarises the proportion of hyperstable regions that are present in each sample, and adds to columns of the sampleTable
}
