% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQC.R
\name{addHyperStableFraction}
\alias{addHyperStableFraction}
\title{Summarise the fraction of hyperstably methylated regions with clear methylation (hg38 only)
This function calculates and summarises the proportion of hyperstable regions that are present in each sample,
and adds this as a column to the sampleTable. Uses a set of windows shown to be methylated across a large number of
human methylation arrays over a wide range of tissues and diseases by \href{https://pubmed.ncbi.nlm.nih.gov/25493099/}{Edgar et al (2014)}}
\usage{
addHyperStableFraction(qseaSet, minDensity = 5, minBeta = 0.8)
}
\arguments{
\item{qseaSet}{A qseaSet object}

\item{minDensity}{A value of \code{CpG_density} to only consider regions above. Defaults to 5.}

\item{minBeta}{The minimum beta value to consider a region as being called as methylated. Defaults to 0.8}
}
\description{
Summarise the fraction of hyperstably methylated regions with clear methylation (hg38 only)
This function calculates and summarises the proportion of hyperstable regions that are present in each sample,
and adds this as a column to the sampleTable. Uses a set of windows shown to be methylated across a large number of
human methylation arrays over a wide range of tissues and diseases by \href{https://pubmed.ncbi.nlm.nih.gov/25493099/}{Edgar et al (2014)}
}
