% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{addLibraryInformation}
\alias{addLibraryInformation}
\title{This function takes a qseaSet and adds the information on the number of reads into the sampleTable for easier access.}
\usage{
addLibraryInformation(qseaSet)
}
\arguments{
\item{qseaSet}{The qseaSet object.}
}
\value{
A qseaSet object with the sampleTable enhanced with the information on number of reads etc
}
\description{
This function takes a qseaSet and adds the information on the number of reads into the sampleTable for easier access.
}
\examples{
#original columns on the sampleTable
colnames(getSampleTable(exampleTumourNormal))
updated <- addLibraryInformation(exampleTumourNormal) 
# now the qseaSet has additional columns from the libraries slot in the qseaSet
colnames(getSampleTable(updated))
}
