% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEnrichment.R
\name{addMedipsEnrichmentFactors}
\alias{addMedipsEnrichmentFactors}
\title{This function takes a qseaSet and adds the Medips-style enrichment scores to the object sampleTable.}
\usage{
addMedipsEnrichmentFactors(
  qseaSet,
  BSgenome = "BSgenome.Hsapiens.NCBI.GRCh38",
  exportPath = NULL,
  extend = 0,
  shift = 0,
  uniq = 0,
  chr.select = NULL,
  paired = TRUE,
  file_name = "file_name",
  nCores = 1
)
}
\arguments{
\item{qseaSet}{The cutoff to use on the windows for each sample}

\item{BSgenome}{Name of BSgenome package}

\item{exportPath}{Folder to export files}

\item{extend, shift, uniq, chr.select, paired}{Options to feed into medips::getGRange or medips::getPairedGRange}

\item{file_name}{Column of the sampleTable which contains the file_name.}

\item{nCores}{Number of cores to use for parallelisation}
}
\value{
A qseaSet supplemented with more columns on the sampleTable, related to enrichment.
}
\description{
This function takes a qseaSet and adds the Medips-style enrichment scores to the object sampleTable.
}
