% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQset.R
\name{addNormalisation}
\alias{addNormalisation}
\title{This function takes a qseaSet and adds several normalisation steps from qsea, with default values.}
\usage{
addNormalisation(
  qseaSet,
  enrichmentMethod = "blind1-15",
  maxPatternDensity = 0.05
)
}
\arguments{
\item{qseaSet}{A qseaSet object.}

\item{enrichmentMethod}{What method to use to calculate the enrichment step.
Currently implemented is the blind1-15, the blind calibration method detailed
in the qsea paper of fitting of a straight line of decreasing expected average
methylation levels from 76\% at CpG_density = 1 to 25\% at CpG_density = 15.}

\item{maxPatternDensity}{Maximum pattern density in a window to consider it for the background calculation.}
}
\value{
A qseaSet object with normalisation steps added.
}
\description{
This function takes a qseaSet and adds several normalisation steps from qsea, with default values.
}
\examples{
getExampleQseaSet(expSamplingDepth = 100000) \%>\% addNormalisation(maxPatternDensity = 0.5)
}
