% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{addSummaryAcrossWindows}
\alias{addSummaryAcrossWindows}
\title{This function takes a qseaSet and adds to the sampleTable summary statistics calculated over a set of windows.}
\usage{
addSummaryAcrossWindows(
  qseaSet,
  regionsToOverlap = NULL,
  fn = mean,
  suffix = "",
  normMethod = c("nrpm", "beta"),
  naMethod = "impute",
  minEnrichment = 3
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{regionsToOverlap}{A GRanges object (or data frame coercible to one) containing the windows to summarise over. If missing, will use all the regions in the qseaSet.}

\item{fn}{A function to apply across the windows. e.g. mean, median, sd.}

\item{suffix}{A suffix for adding to the new columns of data, to clarify where the regions came from for instance.}

\item{normMethod}{One or more normalisation methods to use, e.g. "nrpm" or "beta" or c("nrpm", "beta").}

\item{naMethod}{What method to use to deal with NA values (for beta values). Options are "drop" or "impute".}

\item{minEnrichment}{For beta values only, the minimum number of reads required for a window to be fully methylated, qsea replaces with NA below this value.}
}
\value{
A table of data, one row per sample, with columns indicating the application of the summary statistic to the
}
\description{
This function takes a qseaSet and adds to the sampleTable summary statistics calculated over a set of windows.
}
\examples{
# add to the sampleTable the median NRPM and beta values across all windows
exampleTumourNormal \%>\% addSummaryAcrossWindows(fn = mean) \%>\% getSampleTable()
# maximum NRPM across a smaller region of the genome
rgns <- data.frame(seqnames = 7, start = 25002001, end = 25017900)
exampleTumourNormal \%>\% addSummaryAcrossWindows(regionsToOverlap = rgns, fn = max, normMethod = "nrpm") \%>\% getSampleTable()
# can add a suffix and apply repeatedly with different windows
exampleTumourNormal \%>\% 
  addSummaryAcrossWindows(regionsToOverlap = rgns, fn = median, normMethod = "nrpm", suffix = "subset") \%>\% 
  addSummaryAcrossWindows(fn = median, normMethod = "nrpm", suffix = "all") \%>\% 
  getSampleTable()
}
