% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{filterByOverlaps}
\alias{filterByOverlaps}
\alias{filterByNonOverlaps}
\title{Filter a qseaSet by overlapping/non-overlapping GRanges object or table that can be coerced to one}
\usage{
filterByOverlaps(qseaSet, regionsToOverlap)

filterByNonOverlaps(qseaSet, regionsToOverlap)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{regionsToOverlap}{A set of windows to keep, either a GRanges object, a dataframe with seqnames, start and end.}
}
\value{
A qseaSet object, with only a subset of the windows.
}
\description{
Filter a qseaSet by overlapping/non-overlapping GRanges object or table that can be coerced to one
}
\examples{

rgns <- data.frame(seqnames = 7, start = 25002001, end = 25017900)
filterByOverlaps(exampleTumourNormal, regionsToOverlap = rgns)
filterByNonOverlaps(exampleTumourNormal, regionsToOverlap = rgns)
}
\seealso{
Check the plyranges documentation on \link[plyranges]{filter_by_overlaps} and \link[plyranges]{filter_by_non_overlaps}
}
