% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{arrange.qseaSet}
\alias{arrange.qseaSet}
\title{This function extends the function dplyr::arrange to act on qseaSet, to reorder the samples in the qseaSet.}
\usage{
\method{arrange}{qseaSet}(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A qseaSet to reorder the samples in.}

\item{...}{Other arguments to pass to dplyr::arrange}

\item{.by_group}{Not implemented as qsea requires a data frame not a tibble for the sampleTable.}
}
\value{
A qseaSet with the sample names reordered according to a column of the sampleTable.
}
\description{
This function extends the function dplyr::arrange to act on qseaSet, to reorder the samples in the qseaSet.
}
