% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asValidGranges}
\alias{asValidGranges}
\title{This function checks that an object can be coerced into a GRanges object and does so if possible}
\usage{
asValidGranges(object)
}
\arguments{
\item{object}{An object that may be a GRanges object.}
}
\description{
This function checks that an object can be coerced into a GRanges object and does so if possible
}
