% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEnrichment.R
\name{calculateCGEnrichment}
\alias{calculateCGEnrichment}
\title{This function takes a path to a bam file and calculates the enrichment scores.}
\usage{
calculateCGEnrichment(
  file = NULL,
  BSgenome = NULL,
  exportPath = NULL,
  extend = 0,
  shift = 0,
  uniq = 0,
  chr.select = NULL,
  paired = TRUE
)
}
\arguments{
\item{file}{The path to the bam file}

\item{BSgenome}{Which BSGenome to use (GRCh38 is cached for speed)}

\item{exportPath}{Location to save a plot and data object into. This contains a histogram of fragment length distribution and a rds file containing the}

\item{extend, shift, uniq, chr.select, paired}{Options to feed into MEDIPS::getGRange or MEDIPS::getPairedGRange}
}
\value{
A data frame containing the relH, GoGe and number of reads values for the samples
}
\description{
This function takes a path to a bam file and calculates the enrichment scores.
}
