% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDMRs.R
\name{calculateDMRs}
\alias{calculateDMRs}
\title{Fit the GLM and return the data}
\usage{
calculateDMRs(
  qseaSet,
  variable = NULL,
  sampleNames = NULL,
  covariates = NULL,
  contrastsToDo = NULL,
  minReadCount = 0,
  minNRPM = 1,
  checkPVals = TRUE,
  fdrThres = 0.05,
  keepPvals = FALSE,
  formula = NULL,
  keepData = FALSE,
  keepGroupMeans = FALSE,
  direction = "both"
)
}
\arguments{
\item{qseaSet}{The qseaSet object}

\item{variable}{Which variable to use to calculate the DMRs between}

\item{sampleNames}{Which samples to return the information for.}

\item{covariates}{Any variables to use as covariates, for instance patient in a paired analysis}

\item{contrastsToDo}{A data frame with two columns, sample1 and sample2, with the strings to compare between in each. Multiple rows means that multiple comparisons will be fitted}

\item{minReadCount}{A minimum read count that at least one sample (in the contrasts) must reach in order to consider the region for further calculation. Preferably use minNRPM.}

\item{minNRPM}{A minimum normalised reads per million value that at least one sample (in the contrasts) must reach in order to consider the region for further calculation. Set this or minReadCount (but preferably this).}

\item{checkPVals}{Whether to check that the p-values aren't all zero to avoid a bug with covariates, only turn this off if you are very sure what you are doing!}

\item{fdrThres}{False discovery rate threshold.}

\item{keepPvals}{Whether to keep the unadjusted p-values in the output}

\item{formula}{Alternative formula mode for calculating DMRs (not recommended)}

\item{keepData}{Whether to keep the individual data columns in the output}

\item{keepGroupMeans}{Whether to keep the group means in the output}

\item{direction}{Whether to keep regions that are up/down/both.}
}
\value{
A tibble with the data
}
\description{
This function fits a GLM and returns the resulting data table
}
