% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{calculateFractionReadsInGRanges}
\alias{calculateFractionReadsInGRanges}
\title{This function calculates the fraction of reads in a qseaSet object which are present in a GRanges object}
\usage{
calculateFractionReadsInGRanges(qseaSet, windowsToConsider, numCountsNeeded)
}
\arguments{
\item{qseaSet}{A qseaSet}

\item{windowsToConsider}{Which windows to consider, as a GRanges object}

\item{numCountsNeeded}{Minimum number of reads required to consider that window}
}
\description{
This function calculates the fraction of reads in a qseaSet object which are present in a GRanges object
}
