% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEnrichment.R
\name{calculateGenomicCGDistribution}
\alias{calculateGenomicCGDistribution}
\title{This function takes a BSGenome and returns a data frame with the relH and GoGe parameters used for calculating enrichment}
\usage{
calculateGenomicCGDistribution(BSgenome)
}
\arguments{
\item{BSgenome}{The BSgenome object.}
}
\value{
A data frame with two entries, genome.relH and genome.GoGe
}
\description{
This function takes a BSGenome and returns a data frame with the relH and GoGe parameters used for calculating enrichment
}
