% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{colnames,qseaSet-method}
\alias{colnames,qseaSet-method}
\title{Column names of a qseaSet sample table}
\usage{
\S4method{colnames}{qseaSet}(x)
}
\arguments{
\item{x}{\code{qseaSet}.
The qseaSet object whose \code{sampleTable} column names are to be retrieved.}
}
\value{
\code{character()}.
A character vector containing the column names of the sample table.
}
\description{
S4 method for \code{\link[BiocGenerics]{colnames}} that returns the column names of a
\code{qseaSet}'s sample metadata table (i.e., \code{\link[qsea:qseaSet-class]{qsea::getSampleTable()}}).
}
\examples{
data(exampleTumourNormal, package = "mesa")

# Retrieve the column names of the sample metadata
colnames(exampleTumourNormal)

}
\seealso{
\code{\link[qsea:qseaSet-class]{qsea::getSampleTable()}},
\code{\link[qsea:qseaSet-class]{qsea::getSampleNames()}}
}
