% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineQsets.R
\name{combineQsets}
\alias{combineQsets}
\title{Combine two qseaSets}
\usage{
combineQsets(
  qseaSet1,
  qseaSet2,
  checkParams = FALSE,
  regionsToKeep = NULL,
  dropDuplicates = FALSE
)
}
\arguments{
\item{qseaSet1}{The first qseaSet object.}

\item{qseaSet2}{The second qseaSet object.}

\item{checkParams}{Whether to verify that all the parameters are identical in both objects}

\item{regionsToKeep}{A GRanges object (or table coercible to one) to use to subset the samples. Helps keep RAM use down.}

\item{dropDuplicates}{Whether to drop samples with the same name. Else renames them by adding "_Dup" to the end of the name.}
}
\value{
A qseaSet object, containing all the samples from both qseaSet objects.
}
\description{
This function takes two qseaSets and combines them into one.
}
