% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{convertToArrayBetaTable}
\alias{convertToArrayBetaTable}
\title{This function calculates the beta values for the windows covering probes from a methylation array.}
\usage{
convertToArrayBetaTable(qseaSet, arrayType = "Infinium405k")
}
\arguments{
\item{qseaSet}{qseaSet object to calculate approximate beta values for}

\item{arrayDetails}{Either a}
}
\description{
This function calculates the beta values for the windows covering probes from a methylation array.
}
