% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{downSample}
\alias{downSample}
\title{This function takes a qseaSet and downsamples the reads.}
\usage{
downSample(qseaSet, nReads, renormalise = TRUE)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{nReads}{How many reads to downsample to.}

\item{renormalise}{Whether to renormalise the resulting qseaSet.}
}
\value{
A qseaSet object with the reads downsampled.
}
\description{
This function takes a qseaSet and downsamples the reads.
}
