% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{downSample}
\alias{downSample}
\title{This function takes a qseaSet and downsamples the reads in the counts matrix.}
\usage{
downSample(qseaSet, nReads)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{nReads}{How many reads to downsample to.}
}
\value{
A qseaSet object with the reads downsampled.
}
\description{
This function takes a qseaSet and downsamples the reads in the counts matrix.
}
\examples{
new <- exampleTumourNormal \%>\% downSample(100)
new \%>\% getCountTable() \%>\% dplyr::select(matches("_[NT]")) \%>\% colSums()

}
