% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropAvgFragDetails}
\alias{dropAvgFragDetails}
\title{This function removes the avgFragmentLength and avgFragmentMAPQ columns from the regions of the qseaSet}
\usage{
dropAvgFragDetails(qseaSet)
}
\arguments{
\item{qseaSet}{A qseaSet object to drop the columns from the regions.}
}
\description{
This function removes the avgFragmentLength and avgFragmentMAPQ columns from the regions of the qseaSet
}
