% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePairedTumourQset.R
\docType{data}
\name{examplePairedTumourQset}
\alias{examplePairedTumourQset}
\title{examplePairedTumourQset}
\format{
examplePairedTumourQset
}
\usage{
examplePairedTumourQset
}
\description{
examplePairedTumourQset
}
\keyword{datasets}
