% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleTumourNormal.R
\docType{data}
\name{exampleTumourNormal}
\alias{exampleTumourNormal}
\title{A small example qseaSet with 5 paired tumour/normal samples.}
\format{
A qseaSet object with 819 regions and 10 samples:
}
\usage{
exampleTumourNormal
}
\description{
A qseaSet object containing a small region of chromosome 7 for 5 paired tumour/normal samples (commercially bought from Origene and processed within CRUK-MI CBC).
}
\keyword{datasets}
