% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{filter.qseaSet}
\alias{filter.qseaSet}
\title{This function extends the dplyr function filter to act on qseaSet sampleTable.}
\usage{
\method{filter}{qseaSet}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{...}{Other arguments to pass to dplyr::filter}

\item{.preserve}{Not implemented as irrelevant for qseaSet}

\item{qseaSet}{A qseaSet to filter on}
}
\value{
A qseaSet object with the sampleTable enhanced with the information on number of reads etc
}
\description{
This function extends the dplyr function filter to act on qseaSet sampleTable.
}
