% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{filterWindows}
\alias{filterWindows}
\title{Filter regions (windows) inside a qseaSet}
\usage{
filterWindows(qseaSet, ...)
}
\arguments{
\item{qseaSet}{A \code{qseaSet}.}

\item{...}{Predicates passed to \code{\link[dplyr:filter]{dplyr::filter()}} (applied to the regions).}
}
\value{
The input \code{qseaSet} object with regions filtered.
}
\description{
Filters the \strong{regions} of a \code{qseaSet} using \code{\link[dplyr:filter]{dplyr::filter()}} predicates
applied to its regions as a data frame, then subsets the object accordingly.
}
\examples{
\donttest{
if (system.file("data","exampleTumourNormal.rda",package="mesa") != "") {
  data(exampleTumourNormal, package="mesa")
  # keep short windows as a toy example
  filterWindows(exampleTumourNormal, width < 500)
}
}
}
\seealso{
\code{\link[qsea:qseaSet-class]{qsea::getRegions()}}, \code{\link[=filter.qseaSet]{filter.qseaSet()}}
}
