% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{filterWindows}
\alias{filterWindows}
\title{This function takes a qseaSet object and filters the regions inside it by a call to dplyr::filter.}
\usage{
filterWindows(qseaSet, ...)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{...}{Additional arguments to be used to filter the regions ONLY, as if they were a data frame.}
}
\value{
A qseaSet object, with the regions filtered appropriately.
}
\description{
This function takes a qseaSet object and filters the regions inside it by a call to dplyr::filter.
}
\examples{
exampleTumourNormal \%>\% filterWindows(start <= 25020000)
exampleTumourNormal \%>\% filterWindows(CpG_density > 10)
}
