% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDMRs.R
\name{fitQseaGLM}
\alias{fitQseaGLM}
\title{Fit a generalised linear model}
\usage{
fitQseaGLM(
  qseaSet,
  variable = NULL,
  covariates = NULL,
  contrasts = NULL,
  keepIndex = NULL,
  minReadCount = 0,
  minNRPM = 1,
  checkPVals = TRUE,
  formula = NULL,
  shareDispersionOutsideContrasts = FALSE
)
}
\arguments{
\item{qseaSet}{The qseaSet object}

\item{variable}{The variable to treat as the independent variable between samples in the linear model}

\item{covariates}{A vector including any additional covariates to include in the model}

\item{contrasts}{A data frame with the comparisons to do}

\item{keepIndex}{A vector of indices to keep, overwrites minReadCount.}

\item{minReadCount}{A minimum read count for a row to be considered, in the qseaSet as provided.}

\item{minNRPM}{A minimum normalised reads per million to apply}

\item{checkPVals}{Whether to check excessive numbers of the p-values are exactly zero, to catch a bug in qsea.}

\item{formula}{A formula to use}

\item{shareDispersionOutsideContrasts}{Whether to use samples that are not present in the contrasts to fit the initial generalised linear model, including them in the calculation of dispersion estimates.
Setting this to be TRUE will mean that adding additional samples to the qseaSet will change the calculated DMRs, even if they are not being compared across.}
}
\value{
A qseaGLM object
}
\description{
This function fits a generalised linear model given a qseaSet, and then add a set of contrasts to it
}
