% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get an annotation data frame, for use in heatmap plotting functions}
\usage{
getAnnotation(qseaSet, useGroupMeans = FALSE, sampleAnnotation = NULL)
}
\arguments{
\item{qseaSet}{A qseaSet}

\item{useGroupMeans}{Whether to use the "group" variable to collapse replicates}

\item{sampleAnnotation}{Columns of the sampleTable to use}
}
\value{
A data frame containing the annotation columns, ready for use in heatmap functions
}
\description{
Note that sampleAnnotation must be enclosed within double curly brackets when used.
}
\examples{

# works for none, one or more columns:
exampleTumourNormal \%>\% getAnnotation()
exampleTumourNormal \%>\% getAnnotation(sampleAnnotation = type)
exampleTumourNormal \%>\% getAnnotation(sampleAnnotation = c(tumour,type))
# also works for quoted string names
exampleTumourNormal \%>\% getAnnotation()
exampleTumourNormal \%>\% getAnnotation(sampleAnnotation = "type")
exampleTumourNormal \%>\% getAnnotation(sampleAnnotation = c("tumour","type"))
and for asking to use the groups
exampleTumourNormal \%>\% mutate(group = stringr::str_remove(sample_name,"[0-9]")) \%>\% getAnnotation(useGroupMeans = TRUE)
exampleTumourNormal \%>\% mutate(group = stringr::str_remove(sample_name,"[0-9]")) \%>\% getAnnotation(sampleAnnotation = tumour, useGroupMeans = TRUE)
}
