% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getAnnotationDataFrame}
\alias{getAnnotationDataFrame}
\title{This function generates a data frame to use for the annotation of heatmaps, when using groups}
\usage{
getAnnotationDataFrame(qseaSet, ...)
}
\arguments{
\item{qseaSet}{A qseaSet object}

\item{...}{Which columns of the sampleTable to use in the annotation data frame}
}
\description{
This function generates a data frame to use for the annotation of heatmaps, when using groups
}
