% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQset.R
\name{getBamCoveragePairedAndUnpairedR1}
\alias{getBamCoveragePairedAndUnpairedR1}
\title{This function takes a bam file and gets the coverage over a set of regions. It keeps pairs or high quality R1s.}
\usage{
getBamCoveragePairedAndUnpairedR1(
  fileName = NULL,
  regions = NULL,
  fragmentLength = NULL,
  minMapQual = 30,
  maxInsertSize = 1000,
  minInsertSize = 50,
  minReferenceLength = 30,
  properPairsOnly = FALSE
)
}
\arguments{
\item{fileName}{Path to the bam file}

\item{regions}{GRanges object with the regions to calculate the coverage for.}

\item{fragmentLength}{Length to extend unpaired reads to}

\item{minMapQual}{Minimum mapping quality to include a read (on either end if proper pair)}

\item{maxInsertSize}{Maximum insert size for a proper paired read}

\item{minInsertSize}{Minimum insert size for a proper paired read}

\item{minReferenceLength}{Minimum length on the reference coordinates for a non-paired read to be included}

\item{properPairsOnly}{Whether to only keep proper pairs, else high quality R1s are kept.}
}
\value{
A table
}
\description{
This function takes a bam file and gets the coverage over a set of regions. It keeps pairs or high quality R1s.
}
