% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getBetaTable}
\alias{getBetaTable}
\title{This function takes a qseaSet and makes a table of beta values}
\usage{
getBetaTable(qseaSet, useGroupMeans = FALSE)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{useGroupMeans}{Whether to give means of the group column, rather than individual samples.}
}
\value{
A table of data, one row per window
}
\description{
This function takes a qseaSet and makes a table of beta values
}
