% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEnrichment.R
\name{getCGPositions}
\alias{getCGPositions}
\title{This function takes a GRanges object and calculates the enrichment scores.}
\usage{
getCGPositions(BSgenome, chr.select)
}
\arguments{
\item{BSgenome}{Which BSgenome to use}

\item{chr.select}{Which chromosomes to use}
}
\value{
A data frame containing the relH, GoGe and number of reads values for the samples
}
\description{
This function takes a GRanges object and calculates the enrichment scores.
}
