% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getDataTable}
\alias{getDataTable}
\title{This function takes a qseaSet and generates a table of data containing data for each sample}
\usage{
getDataTable(
  qseaSet,
  normMethod = "nrpm",
  groupMeans = FALSE,
  minEnrichment = 3
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{normMethod}{What normalisation method to use (nrpm or beta)}

\item{groupMeans}{Number of reads required for a fully methylated region to not be masked (put NA) beta values only.}

\item{minEnrichment}{Minimum number of reads for beta values to not give NA}
}
\value{
A table of data, with a column for each individual sample
}
\description{
This function takes a qseaSet and generates a table of data containing data for each sample
}
