% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getCountTable}
\alias{getCountTable}
\alias{getNRPMTable}
\alias{getBetaTable}
\alias{getDataTable}
\title{This function takes a qseaSet and makes a table of counts. A convenience function that wraps getDataTable.}
\usage{
getCountTable(
  qseaSet,
  useGroupMeans = FALSE,
  addMethodSuffix = FALSE,
  verbose = TRUE
)

getNRPMTable(
  qseaSet,
  useGroupMeans = FALSE,
  addMethodSuffix = FALSE,
  verbose = TRUE
)

getBetaTable(
  qseaSet,
  useGroupMeans = FALSE,
  minEnrichment = 3,
  addMethodSuffix = FALSE,
  verbose = TRUE
)

getDataTable(
  qseaSet,
  normMethod = "nrpm",
  useGroupMeans = FALSE,
  minEnrichment = 3,
  addMethodSuffix = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{useGroupMeans}{Whether to use the group column to average over replicates.}

\item{addMethodSuffix}{Whether to include a suffix corresponding to the normalisation method, such as Sample1_beta. This suffix is always present if multiple normalisationMethods are given.}

\item{verbose}{Whether to give a message detailing the table being made.}

\item{minEnrichment}{Minimum number of reads for beta values to not give NA}

\item{normMethod}{What normalisation method to use (nrpm or beta), can be given multiple.}
}
\value{
A table of counts, one row per window

A table of data, one row per window

A table of beta values, one row per window

A table of data, with a column for each individual sample
}
\description{
This function takes a qseaSet and makes a table of counts. A convenience function that wraps getDataTable.

A function takes a qseaSet and makes a table of nrpm values. A convenience function that wraps getDataTable.

A function takes a qseaSet and makes a table of beta values. A convenience function that wraps getDataTable.

This function takes a qseaSet and generates a table of data containing data for each sample
}
\section{Functions}{
\itemize{
\item \code{getCountTable()}: 

\item \code{getNRPMTable()}: 

\item \code{getBetaTable()}: 

}}
