% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getDataTable}
\alias{getDataTable}
\title{This function takes a qseaSet and generates a table of data containing data for each sample}
\usage{
getDataTable(
  qseaSet,
  normMethod = "nrpm",
  useGroupMeans = FALSE,
  minEnrichment = 3,
  addMethodSuffix = FALSE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{normMethod}{What normalisation method to use (nrpm or beta), can be given multiple.}

\item{useGroupMeans}{Whether to use the group column to average over replicates.}

\item{minEnrichment}{Minimum number of reads for beta values to not give NA}

\item{addMethodSuffix}{Whether to include a suffix corresponding to the normalisation method, such as Sample1_beta. This suffix is always present if multiple normalisationMethods are given.}
}
\value{
A table of data, with a column for each individual sample
}
\description{
This function takes a qseaSet and generates a table of data containing data for each sample
}
