% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDMRs.R
\name{getDMRsData}
\alias{getDMRsData}
\title{Extract the data table from all the contrasts}
\usage{
getDMRsData(
  qseaSet,
  qseaGLM,
  sampleNames = NULL,
  variable = NULL,
  keepData = FALSE,
  keepGroupMeans = FALSE,
  fdrThres = 0.05,
  keepPvals = FALSE,
  keepFragmentInfo = FALSE,
  direction = "both"
)
}
\arguments{
\item{qseaSet}{The qseaSet object}

\item{qseaGLM}{A qseaGLM object}

\item{sampleNames}{Which samples to return the information for.}

\item{variable}{Which variable the DMR comparison has been performed on}

\item{keepData}{Whether to keep the individual data columns}

\item{keepGroupMeans}{Whether to keep the group means}

\item{fdrThres}{False discovery rate threshold.}

\item{keepPvals}{Whether to keep the unadjusted p-values in the output}

\item{keepFragmentInfo}{Whether to keep information on the average fragment length and MAPQ (if present)}

\item{direction}{Whether to use regions that are up/down/both.}
}
\value{
A tibble with the data
}
\description{
This function extracts the information from the DMR contrasts into a wide table
}
