% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getGenomicFeatureDistribution}
\alias{getGenomicFeatureDistribution}
\title{This function takes a qseaSet and calculates some stats as to which genomic regions the reads lie.}
\usage{
getGenomicFeatureDistribution(
  qseaSet,
  cutoff = 1,
  normMethod = "nrpm",
  minEnrichment = 3
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{cutoff}{The value required to call a window as being above that cutoff or not}

\item{normMethod}{What normalisation method to use (nrpm or beta)}

\item{minEnrichment}{Number of reads required for a fully methylated region to not be masked (put NA) beta values only.}
}
\value{
A table of summarised data, with windows split out into both the landscape (island/shore/shelf) and genomic region (promoter/exon/intron etc)
}
\description{
This function takes a qseaSet and calculates some stats as to which genomic regions the reads lie.
}
