% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{getMesaGenome}
\alias{getMesaGenome}
\title{Get current mesa genome setting}
\usage{
getMesaGenome()
}
\value{
Character string of current genome setting (e.g., "hg38", "hg19", "mm10")
}
\description{
Get current mesa genome setting
}
