% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{getMesaTxDb}
\alias{getMesaTxDb}
\title{Get TxDb for current or specified genome}
\usage{
getMesaTxDb(genome = NULL)
}
\arguments{
\item{genome}{Genome build, defaults to current setting}
}
\description{
Get TxDb for current or specified genome
}
