% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getNRPMTable}
\alias{getNRPMTable}
\title{This function takes a qseaSet and makes a table of nrpm values}
\usage{
getNRPMTable(qseaSet, useGroupMeans = FALSE)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{useGroupMeans}{Whether to give means of the group column, rather than individual samples.}
}
\value{
A table of data, one row per window
}
\description{
This function takes a qseaSet and makes a table of nrpm values
}
