% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPattern}
\alias{getPattern}
\title{This function returns the name of the pattern added by addPatternDensity, e.g. "CpG"}
\usage{
getPattern(qseaSet)
}
\arguments{
\item{qseaSet}{A qseaSet with pattern details added by addPatternDensity}
}
\description{
This function returns the name of the pattern added by addPatternDensity, e.g. "CpG"
}
\examples{
getPattern(exampleTumourNormal)

}
