% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getSampleGroups2}
\alias{getSampleGroups2}
\title{Sample names per group (order-preserving)}
\usage{
getSampleGroups2(qseaSet)
}
\arguments{
\item{qseaSet}{qsea::qseaSet}
}
\value{
named list of character vectors.
}
\description{
Construct a named list mapping each \code{group} to the vector of \code{sample_name}s
in that group, preserving original ordering.
}
\examples{
data(exampleTumourNormal, package = "mesa")
qs <- exampleTumourNormal
# getSampleGroups2(qs)

}
\seealso{
Other table-helpers: 
\code{\link{getBetaTable}()},
\code{\link{getCountTable}()},
\code{\link{getDataTable}()},
\code{\link{getNRPMTable}()},
\code{\link{makeTransposedTable}()},
\code{\link{removeLibraryFactors}()},
\code{\link{removeNormMethodSuffix}()}
}
\concept{table-helpers}
\keyword{internal}
