% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{getSampleGroups2}
\alias{getSampleGroups2}
\title{This function takes a qseaSet and constructs a list with the groups, but retaining the original sorting.}
\usage{
getSampleGroups2(qseaSet)
}
\arguments{
\item{qseaSet}{The qseaSet object.}
}
\value{
A named list, where each element contains the vector of sample_names that are in that group.
}
\description{
This function takes a qseaSet and constructs a list with the groups, but retaining the original sorting.
}
\examples{
exampleTumourNormal \%>\% mesa:::getSampleGroups2()
}
