% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQC.R
\name{getSampleQCSummary}
\alias{getSampleQCSummary}
\title{This function returns the most important columns of a qseaSet with respect to quality controls}
\usage{
getSampleQCSummary(qseaSet)
}
\arguments{
\item{qseaSet}{A qseaSet object}
}
\description{
This function returns the most important columns of a qseaSet with respect to quality controls
}
\examples{
exampleTumourNormal \%>\% getSampleQCSummary()

}
