% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getWindowAnnotation}
\alias{getWindowAnnotation}
\title{Get an annotation data frame for windows, for use in heatmap plotting functions}
\usage{
getWindowAnnotation(
  dataTab,
  regions,
  windowAnnotation = NULL,
  clusterRows = FALSE
)
}
\arguments{
\item{dataTab}{A table with the data generated from the qseaSet}

\item{regions}{The regions used to generate the annotation for}

\item{windowAnnotation}{Columns of the regions annotation or the dataTab to use}

\item{clusterRows}{Whether the rows are going to be clustered. If not, ensure the annotation is the correct order.}
}
\value{
A data frame containing the annotation columns, ready for use in
}
\description{
Note that windowAnnotation must be enclosed within double curly brackets when used.
}
