% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWindowNames}
\alias{getWindowNames}
\title{This function returns a vector of window names from a qseaSet or a data frame.}
\usage{
getWindowNames(x)
}
\arguments{
\item{x}{A data frame or qseaSet to return the window names of}
}
\description{
This function returns a vector of window names from a qseaSet or a data frame.
}
