% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWindowNames}
\alias{getWindowNames}
\title{This function returns the window names from a data frame with DMR windows}
\usage{
getWindowNames(dataTable)
}
\arguments{
\item{dataTable}{A data frame to return the window names from}
}
\description{
This function returns the window names from a data frame with DMR windows
}
