% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWindows}
\alias{getWindows}
\title{Extract the regions (windows) used in a qseaSet}
\usage{
getWindows(qseaSet)
}
\arguments{
\item{qseaSet}{\code{qseaSet}.
Input object.
\strong{Default:} none (must be supplied).}
}
\value{
A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} of windows.
}
\description{
Convenience wrapper for \code{\link[qsea:qseaSet-class]{qsea::getRegions()}}, returning the genomic windows.
}
\examples{
data(exampleTumourNormal, package = "mesa")
exampleTumourNormal \%>\% getWindows() \%>\% head()

}
\seealso{
\code{\link[qsea:qseaSet-class]{qsea::getRegions()}}, \code{\link[=getWindowNames]{getWindowNames()}}
}
