% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWindows}
\alias{getWindows}
\title{Extract the windows used inside the qseaSet.}
\usage{
getWindows(qseaSet)
}
\arguments{
\item{qseaSet}{A qseaSet}
}
\description{
This function is just a renaming wrapper of the getRegions function from qsea.
}
\examples{
getWindows(exampleTumourNormal)

}
