% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38UltraStableProbes.R
\docType{data}
\name{hg38UltraStableProbes}
\alias{hg38UltraStableProbes}
\title{CpG islands for hg38.}
\format{
A GRanges object with 974 ranges and 1 metadata columns:
\describe{
\item{Probe_ID}{Number of bases in each island}
...
}
}
\usage{
hg38UltraStableProbes
}
\description{
A GRanges object containing ultra-stable methylated probes in hg38 coordinates. List of probes from Edgar et al (2014, PMID: 25493099)
}
\keyword{datasets}
