% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38UltraStableProbes.R
\docType{data}
\name{hg38UltraStableProbes}
\alias{hg38UltraStableProbes}
\title{hg38UltraStableProbes}
\format{
hg38UltraStableProbes
}
\usage{
hg38UltraStableProbes
}
\description{
hg38UltraStableProbes
}
\keyword{datasets}
