% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{left_join.mesaDimRed}
\alias{left_join.mesaDimRed}
\title{Left-join onto the sample table of a mesaDimRed}
\usage{
\method{left_join}{mesaDimRed}(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  keep = NULL,
  ...
)
}
\arguments{
\item{x}{\code{mesaDimRed}
Object whose \code{sampleTable} will be joined.}

\item{y}{\code{data.frame}
Table to join.}

\item{by}{\code{character()} or \code{NULL}
Join columns; see \code{\link[dplyr:mutate-joins]{dplyr::left_join()}}. \strong{Default:} \code{NULL}.}

\item{copy}{\code{logical(1)}
See \code{\link[dplyr:mutate-joins]{dplyr::left_join()}}. \strong{Default:} \code{FALSE}.}

\item{suffix}{\code{character(2)}
Suffixes appended to overlapping non-join column names.
\strong{Default:} \code{c(".x", ".y")}.}

\item{keep}{\code{logical(1)} or \code{NULL}
Retain join keys from both tables. \strong{Default:} \code{NULL}.}

\item{...}{Additional arguments to \code{\link[dplyr:mutate-joins]{dplyr::left_join()}}.}
}
\value{
A \code{mesaDimRed} object:
\itemize{
\item \strong{sampleTable} updated with \code{y} via a left join.
}
}
\description{
Extend \code{\link[dplyr:mutate-joins]{dplyr::left_join()}} to operate on the \code{sampleTable} slot of a
\linkS4class{mesaDimRed} object.
}
\examples{
new <- data.frame(sample_name = c("Colon1_N", "Colon2_N"), foo = c(1,2))

exampleTumourNormal \%>\% 
  getPCA() \%>\%
  left_join(new, by = join_by(sample_name)) \%>\%
  getSampleTable()

}
