% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{left_join.mesaDimRed}
\alias{left_join.mesaDimRed}
\title{Left-join onto the sample table of a mesaDimRed}
\usage{
\method{left_join}{mesaDimRed}(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  keep = NULL,
  ...
)
}
\arguments{
\item{x}{A \linkS4class{mesaDimRed}.}

\item{y}{A data.frame to join.}

\item{by, suffix, copy, keep, ...}{Passed to \code{dplyr::left_join()}.}
}
\value{
A \linkS4class{mesaDimRed} with an updated \code{sampleTable}.
}
\description{
Extends \code{dplyr::left_join()} to operate on the \code{sampleTable} slot.
}
\examples{
st <- data.frame(sample_name = c("A","B"),
                 group = c("X","Y"),
                 row.names = "sample_name")
md <- mesaDimRed(res = list(), sampleTable = st,
                 samples = rownames(st), params = list())
ann <- data.frame(rownameCol = c("A","B"), batch = c(1,2))
md3 <- left_join(md, ann, by = "rownameCol")

}
