% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{left_join.qseaSet}
\alias{left_join.qseaSet}
\title{This function extends the dplyr function left_join to act on qseaSet sampleTable.}
\usage{
\method{left_join}{qseaSet}(qseaSet, ...)
}
\arguments{
\item{qseaSet}{A qseaSet to mutate}

\item{...}{Other arguments to pass to dplyr::mutate}
}
\value{
A qseaSet object with the sampleTable changed by a call to dplyr::left_join
}
\description{
This function extends the dplyr function left_join to act on qseaSet sampleTable.
}
