% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{liftOverHg19}
\alias{liftOverHg19}
\title{This function takes a genomic ranges object or data frame, and lifts over from hg19 to hg38.}
\usage{
liftOverHg19(grOrDf)
}
\arguments{
\item{grOrDf}{Either a genomic ranges object or a data frame (with seqnames column), with or without chr in the seqnames}
}
\value{
A genomic ranges object object with coordinates in hg38.
}
\description{
This function takes a genomic ranges object or data frame, and lifts over from hg19 to hg38.
}
