% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDMRs.R
\name{makeAllContrasts}
\alias{makeAllContrasts}
\title{Generate all possible pairwise contrasts}
\usage{
makeAllContrasts(qseaSet, variable)
}
\arguments{
\item{qseaSet}{A \link[qsea:qseaSet-class]{qsea::qseaSet} object.}

\item{variable}{Character scalar. Column in the sample table.}
}
\value{
A tibble with columns \code{group1} and \code{group2}.
}
\description{
Construct a tibble of all pairwise contrasts between levels of a categorical
variable in the sample table of a \link[qsea:qseaSet-class]{qsea::qseaSet}.
}
\examples{
data(exampleTumourNormal, package = "mesa")
makeAllContrasts(exampleTumourNormal, "type")

}
\seealso{
Other DMR-detection: 
\code{\link{calculateDMRs}()},
\code{\link{fitQseaGLM}()},
\code{\link{getDMRsData}()}
}
\concept{DMR-detection}
