% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDMRs.R
\name{makeAllContrasts}
\alias{makeAllContrasts}
\title{Make a full set of contrasts}
\usage{
makeAllContrasts(qseaSet, variable)
}
\arguments{
\item{qseaSet}{The qseaSet object}

\item{variable}{Which variable to use to calculate the DMRs between}
}
\value{
A tibble with the data
}
\description{
This function fits a GLM and returns the resulting data table
}
