% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDMRs.R
\name{makeAllContrasts}
\alias{makeAllContrasts}
\title{Make a full set of possible contrasts for calculating DMRs on.}
\usage{
makeAllContrasts(qseaSet, variable)
}
\arguments{
\item{qseaSet}{The qseaSet object}

\item{variable}{Which variable to use to calculate the DMRs between}
}
\value{
A tibble with two columns, each row with a pair of contrasts
}
\description{
This function returns a set of all possible contrasts
}
\examples{
exampleTumourNormal \%>\% makeAllContrasts(type)


}
