% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{makeGeneHeatmapRowAnnotation}
\alias{makeGeneHeatmapRowAnnotation}
\title{This function makes a row annotation object for the gene heatmap, which hows information about gene regions and CpG density. It is a helper function for plotGeneHeatmap}
\usage{
makeGeneHeatmapRowAnnotation(rowAnnotationDF)
}
\arguments{
\item{rowAnnotationDF}{A dataframe of variables that provide information about the gene regions, which will be used to annotate windows within the heatmap}
}
\description{
This function makes a row annotation object for the gene heatmap, which hows information about gene regions and CpG density. It is a helper function for plotGeneHeatmap
}
