% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{makeHeatmapAnnotations}
\alias{makeHeatmapAnnotations}
\title{This function makes a pair of HeatmapAnnotation objects, and is a helper function for plotRegionsHeatmap and plotCNV.}
\usage{
makeHeatmapAnnotations(
  qseaSet,
  sampleAnnotation = NULL,
  windowAnnotationDf = NULL,
  useGroupMeans = FALSE,
  specifiedAnnotationColors = NA,
  windowOrientation = "row",
  sampleOrientation = "column"
)
}
\arguments{
\item{qseaSet}{Whether to cluster the rows or not.}

\item{sampleAnnotation}{Which columns to use of the sampleTable.}

\item{windowAnnotationDf}{A data frame with the window annotations. This has been generated by getWindowAnnotation.}

\item{useGroupMeans}{Whether to average over the "group" column of the qseaSet.}

\item{specifiedAnnotationColors}{Allow for overwriting of some of the colours with pre-specified values.}

\item{windowOrientation}{Which orientation the window annotation should be.}

\item{sampleOrientation}{Which orientation the sample annotation should be.}
}
\value{
A list containing two annotation objects
}
\description{
This function makes a pair of HeatmapAnnotation objects, and is a helper function for plotRegionsHeatmap and plotCNV.
}
