% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQset.R
\name{makeQset}
\alias{makeQset}
\title{Make the initial qseaSet object from a set of samples}
\usage{
makeQset(
  sampleTable,
  BSgenome = NULL,
  chrSelect = 1:22,
  windowSize = 300,
  CNVwindowSize = 1e+06,
  fragmentType = NULL,
  fragmentLength = NULL,
  fragmentSD = NULL,
  CNVmethod = "HMMdefault",
  coverageMethod = "PairedAndR1s",
  minMapQual = 10,
  minInsertSize = 70,
  maxInsertSize = 1000,
  minReferenceLength = 30,
  badRegions = NULL,
  properPairsOnly = FALSE,
  hmmCopyGC = NULL,
  hmmCopyMap = NULL,
  maxPatternDensity = 0.05,
  enrichmentMethod = "blind1-15",
  parallel = getMesaParallel()
)
}
\arguments{
\item{sampleTable}{A data frame with the sample information to be passed to qsea.}

\item{BSgenome}{A BSgenome string. See BSgenome::available.genomes() for options.}

\item{chrSelect}{Which chromosomes to use (default 1:22).}

\item{windowSize}{What window size (in bp) to use in the genome (default 300).}

\item{CNVwindowSize}{What window size (in bp) to use in the calculation of CNV (default 1000000).}

\item{fragmentType}{What type of procedure generated the library, Sheared or cfDNA (used to set the average fragment length/SD to predefined defaults)}

\item{fragmentLength}{Average DNA fragment length. Can be set by fragmentType.}

\item{fragmentSD}{Standard deviation of the DNA fragment lengths. Can be set by fragmentType.}

\item{CNVmethod}{Which method to use for calculation of CNV. Options include "HMMdefault" (hmmcopy with default parameters) and "None"}

\item{coverageMethod}{Whether to use custom method for reading coverage in ("PairedAndR1s"), rather than qsea's ("qseaPaired").}

\item{minMapQual}{Minimum MAPQ score for a read to be kept. For PairedAndR1s, will keep if either R1 or R2 meet this cutoff in a properly paired read (if MQ tags are set in the bam file with samtools fixmate).}

\item{minInsertSize}{For paired reads, only keep them if they are above a minimum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.}

\item{maxInsertSize}{For paired reads, only keep them if they are below a maximum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.}

\item{minReferenceLength}{A minimum distance on the genome to keep the read. bwa by default gives 19bp as minimum for a read, which is quite short.}

\item{badRegions}{A GRanges object containing regions to filter out from the result.}

\item{properPairsOnly}{Whether to only keep properly paired reads, or to keep high-quality (MAPQ 30+) unpaired R1s as well. Set to TRUE for size selection.}

\item{hmmCopyGC}{A data frame containing GC content per bin (each with size \code{CNVwindowSize}), only for use with hmmcopy.}

\item{hmmCopyMap}{A data frame containing Mapability content per bin (each with size \code{CNVwindowSize}), only for use with hmmcopy.}

\item{maxPatternDensity}{Maximum pattern density in a window to consider it for the background calculation.}

\item{enrichmentMethod}{What method to use to calculate the enrichment step. Current options are "none" or "blind1-15",
the blind calibration method detailed in the qsea paper of fitting of a straight line of decreasing expected average
methylation levels from 76\% at CpG_density = 1 to 25\% at CpG_density = 15.}

\item{parallel}{Whether to read in files by using each core in parallel.
Control number of calls by calling e.g. setMesaParallel(nCores = 4) beforehand.}
}
\value{
A qseaSet object, containing all the information required.
}
\description{
This function makes the initial qseaSet from a sampleTable that includes the file paths for the bam files and their metadata
}
\examples{
\dontshow{if (requireNamespace("MEDIPSData", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# make a sampleTable (requires `MEDIPSData` package to be installed)
sampleTable <- data.frame(sample_name = c("Normal1","Tumour1"),
  group = c("Normal1","Tumour1"),
  file_name = c(system.file("extdata", "NSCLC_MeDIP_1N_fst_chr_20_21_22.bam", package = "MEDIPSData", mustWork = TRUE),
               system.file("extdata", "NSCLC_MeDIP_1T_fst_chr_20_21_22.bam", package = "MEDIPSData", mustWork = TRUE)))
               
makeQset(sampleTable, 
  BSgenome = "BSgenome.Hsapiens.UCSC.hg19", 
  chrSelect = paste0("chr",20:22), 
  fragmentLength = 200, 
  fragmentSD = 50, 
  CNVmethod = "none")
\dontshow{\}) # examplesIf}
}
