% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQset.R
\name{makeQset}
\alias{makeQset}
\title{Make the initial qseaSet object from a set of samples}
\usage{
makeQset(
  sampleTable,
  BSgenome = NULL,
  chrSelect = 1:22,
  windowSize = 300,
  CNVwindowSize = 1000000,
  fragmentType = NULL,
  fragmentLength = NULL,
  fragmentSD = NULL,
  CNVmethod = "HMMdefault",
  coverageMethod = "PairedAndR1s",
  minMapQual = 10,
  minInsertSize = 70,
  maxInsertSize = 1000,
  minReferenceLength = 30,
  badRegions = NULL,
  properPairsOnly = FALSE
)
}
\arguments{
\item{sampleTable}{A data frame with the sample information to be passed to qsea.}

\item{BSgenome}{A BSgenome string. See BSgenome::available.genomes() for options.}

\item{chrSelect}{Which chromosomes to use (default 1:22).}

\item{windowSize}{What window size (in bp) to use in the genome (default 300).}

\item{CNVwindowSize}{What window size (in bp) to use in the calculation of CNV (default 1000000).}

\item{fragmentType}{What type of procedure generated the library, Sheared or cfDNA (used to set the average fragment length/SD to a default for CBC)}

\item{fragmentLength}{Average DNA fragment length. Can be set by fragmentType.}

\item{fragmentSD}{Standard deviation of the DNA fragment lengths. Can be set by fragmentType.}

\item{CNVmethod}{Which method to use for calculation of CNV. Options include "HMMdefault" (hmmcopy with default parameters) and "None"}

\item{coverageMethod}{Whether to use custom methtools method for reading coverage in (set to PairedAndR1s), rather than qsea's (qseaPaired).}

\item{minMapQual}{Minimum MAPQ score for a read to be kept. For PairedAndR1s, will keep if either R1 or R2 meet this cutoff in a properly paired read (if MQ tags are set in the bam file with samtools fixmate).}

\item{minInsertSize}{For paired reads, only keep them if they are above a minimum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.}

\item{maxInsertSize}{For paired reads, only keep them if they are below a maximum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.}

\item{minReferenceLength}{A minimum distance on the genome to keep the read. bwa by default gives 19bp as minimum for a read, which is quite short.}

\item{badRegions}{A GRanges object containing regions to filter out from the result.}

\item{properPairsOnly}{Whether to only keep properly paired reads, or to keep high-quality (MAPQ 30+) unpaired R1s as well. Set to TRUE for size selection.}
}
\value{
A qseaSet object, containing all the information required.
}
\description{
This function makes the initial qseaSet from a sampleTable that includes the file paths for the bam files and their metadata
}
