% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{makeTransposedTable}
\alias{makeTransposedTable}
\title{This function takes a qseaSet and makes a wide transposed table, suitable for using for machine learning}
\usage{
makeTransposedTable(qseaSet, normMethod = "nrpm", ...)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{normMethod}{What normalisation method to use}

\item{...}{Other columns to add from the sampleTable in the output (along with sample_name). Must specify normMethod if using. Can use everything() for all columns.}
}
\value{
A table
}
\description{
This function takes a qseaSet and makes a wide transposed table, suitable for using for machine learning
}
