% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{makeTransposedTable}
\alias{makeTransposedTable}
\title{Generate a wide data table from a qseaSet}
\usage{
makeTransposedTable(qseaSet, normMethod = "nrpm", ...)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{normMethod}{What normalisation method to use}

\item{...}{Other columns to add from the sampleTable in the output (along with sample_name).
Must specify normMethod if using. Can use everything() for all columns, or leave blank to just have the sample_name (as the first column).}
}
\value{
A table where each row represents one sample and each column represents a genomic window.
}
\description{
This function takes a qseaSet and generates a table of data where each row represents one sample and each column represents a genomic window.
Note that
}
