% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesa-package.R
\docType{package}
\name{mesa-package}
\alias{mesa}
\alias{mesa-package}
\title{mesa: Methylation Enrichment Sequencing Analysis}
\description{
A package for the analysis of methylation enrichment sequencing data (e.g. MBD-seq or MEDIP-seq). This allows for the window-based evaluation of methylation levels (using the 'qsea' package), including functions for determining differentially methylated regions. Many functions are provided for tidyverse style modification of the qseaSet objects originally defined in 'qsea'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cruk-mi/mesa}
  \item Report bugs at \url{https://github.com/cruk-mi/mesa}
}

}
\author{
\strong{Maintainer}: Steven Hill \email{stevenhill.work@gmail.com} (\href{https://orcid.org/0000-0002-5909-692X}{ORCID})

Authors:
\itemize{
  \item Simon Pearce (\href{https://orcid.org/0000-0002-1680-5538}{ORCID})
}

Other contributors:
\itemize{
  \item Paddy Harker (\href{https://orcid.org/0000-0002-2531-8774}{ORCID}) [contributor]
  \item Kevin Brennan \email{kevin.brennan@cruk.manchester.ac.uk} (\href{https://orcid.org/0000-0001-9018-9852}{ORCID}) [contributor]
  \item Katarzyna Kamieniecka [contributor]
  \item Felipe Perez Martinez \email{felipe.perezmartinez@cruk.manchester.ac.uk} (\href{https://orcid.org/0009-0007-9138-3674}{ORCID}) [contributor]
  \item Cancer Research UK National Biomarker Centre [copyright holder, funder]
}

}
\keyword{internal}
