% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesa.R
\name{mesa}
\alias{mesa}
\title{A mesa table}
\usage{
mesa(data, ssrOptions = mesaSSROptions, styleOptions = mesaStyleOptions)
}
\arguments{
\item{data}{The data to render.}

\item{ssrOptions}{Server-side rendering options. See \link{mesaSSROptions} for details.}

\item{styleOptions}{Options for styling the table. See \link{mesaStyleOptions} for details.}
}
\description{
Use inside \link{renderMesa} to render a \code{mesa} table in your app.
}
