% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mesaDimRed-class}
\alias{mesaDimRed-class}
\title{The MesaDimRed class}
\description{
This is a class for containing the results of a dimensionality reduction on a qseaSet
}
\section{Slots}{

\describe{
\item{\code{res}}{The results, a list containing individual PCA/UMAP objects.}

\item{\code{sampleTable}}{A sampleTable from the qseaSet.}

\item{\code{samples}}{A character vector of the samples used.}

\item{\code{params}}{The parameters used during construction.}

\item{\code{dataTable}}{The data used to construct the PCA/UMAP object (optional).}
}}

