% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mesaPCA-class}
\alias{mesaPCA-class}
\alias{mesaPCA}
\alias{show,mesaPCA-method}
\title{PCA results container}
\usage{
mesaPCA(prcomp, windows)

\S4method{show}{mesaPCA}(object)
}
\arguments{
\item{prcomp}{\code{prcomp}
PCA fit from \code{\link[stats:prcomp]{stats::prcomp()}}. \strong{Default:} none.}

\item{windows}{\code{character()}
Window IDs used. \strong{Default:} none.}

\item{object}{\code{mesaPCA}}
}
\value{
A \linkS4class{mesaPCA} object containing the PCA fit and window IDs.
}
\description{
Stores a PCA fit (from \code{\link[stats:prcomp]{stats::prcomp()}}) computed over methylation windows.
}
\section{Slots}{

\describe{
\item{\code{prcomp}}{\code{prcomp}
A PCA fit returned by \code{\link[stats:prcomp]{stats::prcomp()}}.}

\item{\code{windows}}{\code{character()}
Window IDs used in the PCA.}
}}

\examples{
set.seed(1)
x <- matrix(rnorm(20), nrow = 5, ncol = 4,
            dimnames = list(paste0("S",1:5), paste0("W",1:4)))
pc <- stats::prcomp(x, center = TRUE, scale. = FALSE)
mp <- mesaPCA(prcomp = pc, windows = colnames(x))
mp

}
\seealso{
\linkS4class{mesaDimRed}
}
