% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mesaPCA-class}
\alias{mesaPCA-class}
\title{The mesaPCA class}
\description{
This is a class for containing a PCA object calculated on a qseaSet
}
\section{Slots}{

\describe{
\item{\code{prcomp}}{A principal component object output from \code{\link[stats:prcomp]{stats::prcomp()}}}

\item{\code{windows}}{A character vector of the windows used.}
}}

