% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mesaUMAP-class}
\alias{mesaUMAP-class}
\alias{mesaUMAP}
\alias{show,mesaUMAP-method}
\title{UMAP results container}
\usage{
mesaUMAP(points, windows)

\S4method{show}{mesaUMAP}(object)
}
\arguments{
\item{points}{A data.frame with one row per sample with the positions in UMAP space.
Row names should be sample IDs.}

\item{windows}{Character vector of window IDs used.}

\item{object}{A \linkS4class{mesaUMAP} object.}
}
\value{
A \linkS4class{mesaUMAP} object.
}
\description{
Stores per-sample coordinates from a UMAP embedding computed over methylation
windows.
}
\section{Slots}{

\describe{
\item{\code{points}}{A data.frame with one row per sample and UMAP coordinates
(e.g., columns \code{UMAP1}, \code{UMAP2}). Row names are sample IDs.}

\item{\code{windows}}{Character vector of window IDs used to compute the embedding.}
}}

\examples{
pts <- data.frame(UMAP1 = c(0.1, -0.2, 0.0),
                  UMAP2 = c(0.3,  0.1, -0.1))
rownames(pts) <- paste0("S", 1:3)
mu <- mesaUMAP(points = pts, windows = c("w1","w2","w3"))
mu

}
\seealso{
\linkS4class{mesaDimRed}
}
