% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mesaUMAP-class}
\alias{mesaUMAP-class}
\title{The mesaUMAP class}
\description{
This is a class for containing UMAP objects calculated on a qseaSet
}
\section{Slots}{

\describe{
\item{\code{points}}{Data frame containing one row for each sample, with the positions in UMAP space.}

\item{\code{windows}}{A character vector of the windows used.}
}}

