% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixQsets.R
\name{mixSamples}
\alias{mixSamples}
\title{This function takes a qseaSet and makes a new sample by mixing two samples}
\usage{
mixSamples(
  qseaSet,
  sample1,
  sample2,
  nReadsTotal,
  proportion,
  newName = NULL,
  groupName = NULL,
  onlyNew = FALSE,
  renormalise = TRUE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{sample1}{First sample name, from which to take proportion of samples}

\item{sample2}{Second sample name}

\item{nReadsTotal}{Number of reads in total to have after mixing}

\item{proportion}{The proportion to take from sample1, the rest will come from sample2}

\item{newName}{A name to give the new sample}

\item{groupName}{A name to use in the group column in the sampleTable}

\item{onlyNew}{Whether to only return the new sample.}

\item{renormalise}{Whether to renormalise the result. Speeds up the process when you are repeatedly subsampling, only need to do it once at the end.}
}
\value{
A qseaSet object with an extra
}
\description{
This function takes a qseaSet and makes a new sample by mixing two samples
}
