% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixQsets.R
\name{mixSamples}
\alias{mixSamples}
\title{This function takes a qseaSet and makes a new sample by mixing two samples.}
\usage{
mixSamples(
  qseaSet,
  sample1,
  sample2,
  nReadsTotal,
  proportion,
  newName = NULL,
  groupName = NULL,
  onlyNew = FALSE,
  renormalise = TRUE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{sample1}{First sample name, from which to take \code{proportion} of reads.}

\item{sample2}{Second sample name.}

\item{nReadsTotal}{Number of reads in total to have after mixing, potentially adjusted if the regions have been filtered.}

\item{proportion}{The proportion to take from sample1, the rest will come from sample2.}

\item{newName}{A name to give the new sample.}

\item{groupName}{A name to use in the group column in the sampleTable.}

\item{onlyNew}{Whether to only return the new sample, or all the previous samples as well.}

\item{renormalise}{Whether to renormalise the result. Speeds up the process when you are repeatedly subsampling, only need to do it once at the end.}
}
\value{
A qseaSet object containing a new sample, potentially with all the previous samples.
}
\description{
Perform in silico mixing between two samples, enabling the generation of new synthetic data.
Note that if the qseaSets have been filtered prior to using this, the fraction of fragments
present in the current qseaSet is calculated, such that the total number of reads would have been approximately \code{nReadsTotal}.
E.g. if there are 50\% of the total fragments in the windows present in the current samples,
0.5*\code{nReadsTotal} fragments will be sampled.
}
\examples{
# give an approximate set of reads as if Colon1_T and Colon1_N were mixed together and sequenced to 100000 total fragments
mixed <- exampleTumourNormal \%>\% mixSamples("Colon1_T", "Colon1_N", nReadsTotal = 100000, proportion = 0.5, renormalise = FALSE)
# only a few reads are included in this small subset of windows:
mixed \%>\% getCountTable() \%>\% pull(Mix_Colon1_T_Colon1_N_0.5) \%>\% sum()
}
