% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mutate.mesaDimRed}
\alias{mutate.mesaDimRed}
\title{Mutate the sample table of a mesaDimRed}
\usage{
\method{mutate}{mesaDimRed}(.data, ...)
}
\arguments{
\item{.data}{\code{mesaDimRed}
Object to modify.}

\item{...}{Arguments passed to \code{\link[dplyr:mutate]{dplyr::mutate()}}.}
}
\value{
A \code{mesaDimRed} object:
\itemize{
\item \strong{sampleTable} updated with the mutated columns.
\item Sample identity is preserved; attempts to alter \code{sample_name} are rejected.
}
}
\description{
Extend \code{\link[dplyr:mutate]{dplyr::mutate()}} to operate on the \code{sampleTable} slot of a
\linkS4class{mesaDimRed} object.
}
\examples{
st <- data.frame(sample_name = c("A","B"),
                 group = c("X","Y"),
                 row.names = "sample_name")
md <- mesaDimRed(res = list(), sampleTable = st,
                 samples = rownames(st), params = list())
md2 <- mutate(md, group2 = paste0(group, "_2"))
stopifnot("group2" \%in\% colnames(
  methods::slot(md2, "sampleTable")
))

}
