% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mutate.mesaDimRed}
\alias{mutate.mesaDimRed}
\title{This function extends the dplyr function mutate to act on a mesaDimRed sampleTable.}
\usage{
\method{mutate}{mesaDimRed}(.data, ...)
}
\arguments{
\item{.data}{A mesaDimRed to mutate}

\item{...}{Other arguments to pass to dplyr::mutate}
}
\value{
A mesaDimRed object with the sampleTable changed by a call to dplyr::mutate
}
\description{
This function extends the dplyr function mutate to act on a mesaDimRed sampleTable.
}
