% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{mutate.qseaSet}
\alias{mutate.qseaSet}
\title{This function extends the dplyr function mutate to act on qseaSet sampleTable.}
\usage{
\method{mutate}{qseaSet}(.data, ...)
}
\arguments{
\item{.data}{A qseaSet to mutate}

\item{...}{Other arguments to pass to dplyr::mutate}
}
\value{
A qseaSet object with the sampleTable changed by a call to dplyr::mutate
}
\description{
This function extends the dplyr function mutate to act on qseaSet sampleTable.
}
