% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\examples{
# Basic usage: pipe a vector into sum()
1:5 \%>\% sum()

# Chain multiple transformations
1:10 \%>\%
  mean() \%>\%
  round(1)
}
\keyword{internal}
