% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseDMRs.R
\name{pivotDMRsLonger}
\alias{pivotDMRsLonger}
\title{This function transforms a DMR data frame from the wide format to a long format}
\usage{
pivotDMRsLonger(DMRtable, FDRthres = 0.05, makePositive = FALSE)
}
\arguments{
\item{DMRtable}{A data frame with multiple comparisons inside}

\item{FDRthres}{FDR threshold to apply to each comparison}

\item{makePositive}{Whether to reverse the contrast when the window is hypomethylated in the contrast}
}
\description{
This function transforms a DMR data frame from the wide format to a long format
}
