% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaCNV.R
\name{plotCNVheatmap}
\alias{plotCNVheatmap}
\title{This function takes a qseaSet and plots a heatmap of the calculated CNV}
\usage{
plotCNVheatmap(qseaSet, sampleAnnotation = NULL, clusterRows = TRUE)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{sampleAnnotation}{Columns of the sampleTable to use to annotation the plot with.}

\item{clusterRows}{Whether to cluster the rows of the heatmap}
}
\value{
A heatmap with the calculated number of chromosomes for each samples
}
\description{
This function takes a qseaSet and plots a heatmap of the calculated CNV
}
