% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{This function takes a qseaSet and plots a correlation matrix}
\usage{
plotCorrelationMatrix(qseaSet, ..., normMethod = "nrpm", annotationColors = NA)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{...}{Other arguments for pheatmap}

\item{normMethod}{What normalisation method to use}

\item{annoDF}{A data frame for use of annotation of the rows and columns in pheatmap}
}
\value{
A table
}
\description{
This function takes a qseaSet and plots a correlation matrix
}
