% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{This function takes a qseaSet and plots a correlation matrix}
\usage{
plotCorrelationMatrix(
  qseaSet,
  regionsToOverlap = NULL,
  useGroupMeans = FALSE,
  sampleAnnotation = NULL,
  normMethod = "nrpm",
  minEnrichment = 3,
  annotationColors = NA,
  minDensity = 0,
  ...
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{regionsToOverlap}{Regions to overlap}

\item{useGroupMeans}{Whether to use the replicate information or not (group column)}

\item{sampleAnnotation}{Columns of the sampleTable to use to annotate the heatmap}

\item{normMethod}{What normalisation method to use}

\item{minEnrichment}{Minimum number of reads for beta values}

\item{annotationColors}{A list of custom colours to use for the annotation}

\item{minDensity}{Minimum CpG density to keep}

\item{...}{Other arguments for pheatmap}
}
\value{
A table
}
\description{
This function takes a qseaSet and plots a correlation matrix
}
