% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDMRUpset}
\alias{plotDMRUpset}
\title{This function takes a DMRtable as output by calculateDMRs (possibly after any filtering), and generates an upset plot}
\usage{
plotDMRUpset(DMRtable, string = NULL, removeVS = TRUE, minAdjPval = 0.05, ...)
}
\arguments{
\item{DMRtable}{The data frame containing the differentially methylated regions.}

\item{string}{A string to subset the columns based on (always requires adjPval)}

\item{removeVS}{Whether to remove the string "\emph{vs}" and everything after from the column names. For instance if all comparisons are against the same sample(s)}

\item{minAdjPval}{A minimum adjusted P value to consider the window as present}

\item{...}{Other arguments to be passed to upset}
}
\value{
A table
}
\description{
This function takes a DMRtable as output by calculateDMRs (possibly after any filtering), and generates an upset plot
}
