% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGeneHeatmap}
\alias{plotGeneHeatmap}
\title{This function takes a qseaSet and a gene, and plots the expression across the gene as a heatmap}
\usage{
plotGeneHeatmap(
  qseaSet,
  gene,
  normMethod = "beta",
  useGroupMeans = FALSE,
  sampleAnnotation = NULL,
  minDensity = 0,
  minEnrichment = 3,
  maxScale = 1,
  clusterNum = 2,
  annotationColors = NULL,
  upstreamDist = 3000,
  scaleRows = FALSE,
  clusterCols = TRUE,
  mart = NULL,
  downstreamDist = 1000,
  ...
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{gene}{A gene to plot. Either a gene symbol or an ensembl ID.}

\item{normMethod}{Whether to plot nrpm values or beta values.}

\item{useGroupMeans}{Whether to use the group variable to collapse replicates.}

\item{sampleAnnotation}{A data frame with annotations for the samples.}

\item{minDensity}{A minimum CpG density level to filter out windows with values lower than.}

\item{minEnrichment}{Minimum enrichment factor for beta values, will give NAs below this.}

\item{maxScale}{The maximum of the scale, not used when plotting beta values.}

\item{clusterNum}{A number of clusters to break the column dendrogram into.}

\item{annotationColors}{A list with the colours to use for the column legend, to pass to pheatmap}

\item{upstreamDist}{Number of basepairs upstream of the gene to include.}

\item{scaleRows}{Whether to scale the rows of the heatmap}

\item{clusterCols}{Whether to cluster the columns or not.}

\item{mart}{A biomaRt mart object. If not supplied, will check the qseaSet, else will get a default for GRCh38/hg38 or hg19.}

\item{downstreamDist}{Number of basepairs downstream of the gene to include.}

\item{...}{Additional arguments to pass to pheatmap.}
}
\value{
A heatmap showing the methylation patterns across the gene of interest.
}
\description{
This function takes a qseaSet and a gene, and plots the expression across the gene as a heatmap
}
