% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGenomicFeatureDistribution}
\alias{plotGenomicFeatureDistribution}
\title{This function takes a qseaSet and plots the distribution of the genomic features of all the windows over a cutoff}
\usage{
plotGenomicFeatureDistribution(
  qseaSet,
  cutoff = 1,
  barType = "stack",
  normMethod = "nrpm"
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{cutoff}{The cutoff to use on the windows for each sample}

\item{barType}{What type of bars to use for the plot (stack, dodge, fill)}

\item{normMethod}{Normalisation method to use}
}
\value{
A plot of the distribution
}
\description{
This function takes a qseaSet and plots the distribution of the genomic features of all the windows over a cutoff
}
\examples{
exampleTumourNormal \%>\% plotGenomicFeatureDistribution(normMethod = "beta", cutoff = 0.75)

}
