% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plotPCA}
\alias{plotPCA}
\title{This function takes the output of \code{\link[=getPCA]{getPCA()}} and produces plots.}
\usage{
plotPCA(
  object,
  qseaSet = NULL,
  components = list(c(1, 2), c(2, 3)),
  colour = NULL,
  colourPalette = NULL,
  NAcolour = "grey50",
  symDivColourScale = FALSE,
  shape = NULL,
  shapePalette = NULL,
  NAshape = NULL,
  showSampleNames = FALSE,
  pointSize = 2,
  alpha = 1,
  plotlyAnnotations = ""
)
}
\arguments{
\item{object}{The output from \code{\link[=getPCA]{getPCA()}}.}

\item{qseaSet}{The qseaSet object used to generate \code{object}.}

\item{components}{Vector of the two components to plot, or a list of vectors to make multiple plots. Default is to produce plots for PC1 vs PC2 and PC2 vs PC3 for PCA and UMAP1 vs UMAP2 for UMAP.}

\item{colour}{Character vector of variable names from the qseaSet sample table for setting the colour of the points (samples). Separate plots are made for each variable.}

\item{colourPalette}{Character vector giving the colour palette to use for the points (samples). Defaults are used if not supplied.}

\item{NAcolour}{Colour to use for NA values in the \code{colour} variable. Default is "grey50".}

\item{symDivColourScale}{Logical indicating if a diverging colour scale should be symmetric around zero (ignored if a diverging colour scale is not used).}

\item{shape}{Character giving variable name from the qseaSet sample table for setting the shape of the points (samples). Can only accept a single variable name.}

\item{shapePalette}{Shapes to use for the points (samples) for each category of the \code{shape} variable. Can be one of the following options:
\itemize{
\item A numeric vector specifying the set of shapes. Can be either integers between 0 and 20 (line or filled shapes) or integers between 21 and 25 (filled shapes with a border; border colour is set to black).
\item A character specifying which types of shapes to use (with the exact set of shapes set internally by the function). Either:
\itemize{
\item "line-first" (15 line shapes, then 4 filled shapes; max. 19 categories).
\item "filled-first" (4 filled shapes, then 15 line shapes; max. 19 categories).
\item "mixture" (mixture of line and filled shapes; max. 19 categories).
\item "filled+border" (max. 5 categories, or 4 if there are NAs in the \code{shape} variable).
}
\item NULL; defaults are used which is the "mixture" set of shapes for non-diverging colour scales (or no colour scale) and "filled+border" for diverging colour scales.
}}

\item{NAshape}{Shape to use for NA values in the \code{shape} variable. Default is shape 7, or shape 25 if filled shapes with a border are being used.}

\item{showSampleNames}{Logical indicating whether to show the sample names.}

\item{pointSize}{Numeric value to set the size of the points. Default is 2.}

\item{alpha}{Numeric value between 0 and 1 to set alpha of points. Default is 1.}

\item{plotlyAnnotations}{Vector of columns to annotate for plotly, e.g. c("group","tissue")}
}
\value{
A list of ggplot objects: one for each combination of \code{object@res}, \code{colour} and \code{components}.
}
\description{
This function takes the output of \code{\link[=getPCA]{getPCA()}} and produces plots.
}
\examples{
exampleTumourNormal \%>\% getPCA() \%>\% plotPCA()
}
