% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRegionsHeatmap}
\alias{plotRegionsHeatmap}
\title{This function takes a qseaSet and a GRanges object, and plots the expression across the gene as a heatmap}
\usage{
plotRegionsHeatmap(
  qseaSet,
  regionsToOverlap,
  normMethod = "beta",
  sampleAnnotation = NULL,
  annotationColors = NA,
  useGroups = FALSE,
  clusterRows = FALSE,
  clusterCols = TRUE,
  minEnrichment = 3,
  maxScale = 5,
  clusterNum = 2,
  description = "",
  clip = 1e+09,
  minDensity = 0,
  clusterMethod = "ward.D2",
  ...
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{regionsToOverlap}{A genomic ranges to plot.}

\item{normMethod}{Whether to plot nrpm values or beta values.}

\item{sampleAnnotation}{A character vector with names of columns from the sampleTable to use as annotation, or a pre-defined annotation data frame.}

\item{annotationColors}{A list specifying some or all of the colours to use for the annotations.}

\item{useGroups}{Whether to average samples over the group column (i.e. combine replicates)}

\item{clusterRows}{Whether to cluster the rows or not.}

\item{clusterCols}{Whether to cluster the columns or not.}

\item{minEnrichment}{Minimum enrichment factor for beta values, will give NAs below this.}

\item{maxScale}{The maximum of the scale, not used when plotting beta values.}

\item{clusterNum}{A number of clusters to break the column dendrogram into.}

\item{description}{A string to include as part of the title.}

\item{clip}{An upper level to clip the data at, anything higher than this is replaced with the clip value.}

\item{minDensity}{A minimum CpG density level to filter out windows with values lower than.}

\item{clusterMethod}{What method to use to cluster the dendrograms.}

\item{...}{Other arguments to pass to pheatmap.}
}
\value{
A qseaSet object with the sampleTable enhanced with the information on number of reads etc
}
\description{
This function takes a qseaSet and a GRanges object, and plots the expression across the gene as a heatmap
}
