% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{poolSamples}
\alias{poolSamples}
\title{This function takes a qseaSet and merges samples together into a single sample.}
\usage{
poolSamples(qseaSet, mergeString)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{mergeString}{A string to merge on}
}
\value{
A qseaSet object with the samples merged together.
}
\description{
This function takes a qseaSet and merges samples together into a single sample.
}
