% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{pull.qseaSet}
\alias{pull.qseaSet}
\title{Pull a column from a qseaSet sample table}
\usage{
\method{pull}{qseaSet}(.data, var = -1, name = NULL, ...)
}
\arguments{
\item{.data}{A \code{qseaSet}.}

\item{var}{Column to extract (name or position). Defaults to last column.}

\item{name}{Optional column to use for names in the returned vector.}

\item{...}{Passed to \code{\link[dplyr:pull]{dplyr::pull()}}.}
}
\value{
A vector with length equal to the number of samples in the qseaSet.
}
\description{
Extends \code{\link[dplyr:pull]{dplyr::pull()}} to extract a column from the \code{sampleTable} of a \code{qseaSet}.
}
\examples{
\donttest{
if (system.file("data","exampleTumourNormal.rda",package="mesa") != "") {
  data(exampleTumourNormal, package="mesa")
  exampleTumourNormal \%>\% pull(group)
}
}
}
\seealso{
\code{\link[qsea:qseaSet-class]{qsea::getSampleTable()}}, \code{\link[dplyr:pull]{dplyr::pull()}}
}
