% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{pull.qseaSet}
\alias{pull.qseaSet}
\title{This function extends the dplyr function pull to act on qseaSet sampleTable. It returns a column of the sampleTable.}
\usage{
\method{pull}{qseaSet}(.data, var = -1, name = NULL, ...)
}
\arguments{
\item{.data}{A qseaSet.}

\item{var}{A variable specified as a column name or an integer (negative counting from right)}

\item{name}{An optional parameter that specifies the column to be used as names for a named vector. Specified in a similar manner as var.}

\item{...}{Other arguments to pass to dplyr::pull (column name to extract)}
}
\value{
A vector the same size as the number of samples in the qseaSet.
}
\description{
This function extends the dplyr function pull to act on qseaSet sampleTable. It returns a column of the sampleTable.
}
