% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qseaTableToChrGRanges}
\alias{qseaTableToChrGRanges}
\title{Convert a table from makeTable into a GRanges object}
\usage{
qseaTableToChrGRanges(dataTable)
}
\arguments{
\item{dataTable}{The data frame, generally from qsea::makeTable. Can also be a GRanges object.}
}
\value{
A GenomicRanges object
}
\description{
This function adds the column names to be able to make the makeTable output into a hg38 GRanges object.
Basically it converts chr to seqnames, window_start to start, window_end to end, then turns it into a GRanges.
}
