% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_almost_empty_cols}
\alias{remove_almost_empty_cols}
\title{This function removes columns that are mostly NA (with some proportion)
Based off the janitor::remove_empty_cols function.}
\usage{
remove_almost_empty_cols(dat, prop)
}
\arguments{
\item{dat}{A data frame to filter almost empty columns from}

\item{prop}{A proportion of NAs in each column to allow}
}
\description{
This function removes columns that are mostly NA (with some proportion)
Based off the janitor::remove_empty_cols function.
}
