% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaCNV.R
\name{removeCNV}
\alias{removeCNV}
\title{This function takes a qseaSet and removes the CNV data.}
\usage{
removeCNV(qseaSet)
}
\arguments{
\item{qseaSet}{The qseaSet object.}
}
\value{
A qseaSet with the CNV data removed
}
\description{
This function takes a qseaSet and removes the CNV data.
}
\examples{
new <- exampleTumourNormal \%>\% removeCNV()
new \%>\% getCNV() \%>\% head()
}
