% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{removeLibraryFactors}
\alias{removeLibraryFactors}
\title{This function takes a qseaSet and sets all the library factor columns to one, including in the sampleTable.}
\usage{
removeLibraryFactors(qseaSet)
}
\arguments{
\item{qseaSet}{The qseaSet object.}
}
\value{
A table of data, with a column for each individual sample
}
\description{
This function takes a qseaSet and sets all the library factor columns to one, including in the sampleTable.
}
