% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{removeLibraryFactors}
\alias{removeLibraryFactors}
\title{This function takes a qseaSet and sets all the library factor parts to one, including in the sampleTable if present.}
\usage{
removeLibraryFactors(qseaSet)
}
\arguments{
\item{qseaSet}{The qseaSet object.}
}
\value{
A qseaSet, without the library_factor column for each individual sample
}
\description{
This function takes a qseaSet and sets all the library factor parts to one, including in the sampleTable if present.
}
\examples{
exampleTumourNormal \%>\% removeLibraryFactors() \%>\% addLibraryInformation() \%>\% pull(library_factor)
# different to previously set library factors:
exampleTumourNormal  \%>\% addLibraryInformation() \%>\% pull(library_factor)
}
