% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{removeNormMethodSuffix}
\alias{removeNormMethodSuffix}
\title{This function removes the normMethod suffix, if present, from the sample (or sample group) column names in a data frame of normalised values}
\usage{
removeNormMethodSuffix(dataTable, normMethod = NULL)
}
\arguments{
\item{dataTable}{A data frame of normalised values for a set of windows (rows) and samples (columns), e.g. from \code{\link[=getDataTable]{getDataTable()}}.}

\item{normMethod}{A character giving the normalisation method used (e.g. "beta" or "nrpm").}
}
\value{
The dataTable with the normMethod suffix "_{normMethod}" removed from sample (or sample group) column names.
}
\description{
This function removes the normMethod suffix, if present, from the sample (or sample group) column names in a data frame of normalised values
}
\examples{
exampleTumourNormal \%>\% getDataTable(addMethodSuffix = TRUE)  \%>\% removeNormMethodSuffix(normMethod = "nrpm")

}
