% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{renameQsetNames}
\alias{renameQsetNames}
\title{Rename samples in a qseaSet.}
\usage{
renameQsetNames(qseaSet, pattern, replacement = "")
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{pattern}{Pattern to replace in the names of the samples.}

\item{replacement}{Pattern to replace with in the names of the samples.}
}
\value{
A qseaSet object, containing all the samples from both qseaSet objects.
}
\description{
This function takes a qseaSet and renames the samples in it.
}
