% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{renameSamples}
\alias{renameSamples}
\title{Rename samples using a column from the sample table}
\usage{
renameSamples(qseaSet, newNameColumn)
}
\arguments{
\item{qseaSet}{\code{qseaSet}
A qseaSet object containing methylation-enriched sequencing data.}

\item{newNameColumn}{\code{character(1)}
The name of a column in the \code{sampleTable} containing new sample names.
Values must be unique and contain no \code{NA}s.
(Unquoted column names are supported via tidy evaluation.)}
}
\value{
A \code{qseaSet} object:
\itemize{
\item \strong{renameSamples()}: returns the input \code{qseaSet} with sample names replaced
by values from \code{newNameColumn}. All associated slots are updated to reflect
the new naming.
}
}
\description{
Replace sample names with values from a user-supplied column in the
\code{sampleTable} of a \code{qseaSet}, updating all dependent slots consistently.
}
\examples{
data(exampleTumourNormal, package = "mesa")

# Add a column with synthetic names and rename samples based on that column
exampleTumourNormal \%>\%
  dplyr::mutate(newName = paste0("Sample", 
  seq_len(nrow(qsea::getSampleTable(exampleTumourNormal))))) \%>\% 
  renameSamples(newNameColumn = newName) 

}
