% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterQset.R
\name{renameSamples}
\alias{renameSamples}
\title{This function takes a qseaSet and renames the samples based on a column of the sampleTable.}
\usage{
renameSamples(qseaSet, newNameColumn)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{newNameColumn}{A column of the qseaSet to use to rename each sample with.}
}
\value{
A qseaSet object with the samples renamed.
}
\description{
This function takes a qseaSet and renames the samples based on a column of the sampleTable.
}
