% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaCNV.R
\name{runHMMCopy}
\alias{runHMMCopy}
\title{Function to call HMMcopy on the individual reads for a sample}
\usage{
runHMMCopy(CNV_RegionsWithReads, colname, plotDir = NULL)
}
\arguments{
\item{CNV_RegionsWithReads}{Granges object with reads columns}

\item{colname}{Which column of the sample to use}

\item{plotDir}{A directory to export individual HMMcopy plots to.}
}
\value{
A GRanges object with the result of calling HMMcopy.
}
\description{
This calculates and adds CNV to a qseaSet based on just using HMMcopy on each sample, with default parameters.
Assumes gc and map columns have already been added.
}
