% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{select.qseaSet}
\alias{select.qseaSet}
\title{This function extends the dplyr function select to act on qseaSet sampleTable. Can also be used to rename columns.}
\usage{
\method{select}{qseaSet}(.data, ...)
}
\arguments{
\item{.data}{A qseaSet to select columns of the sampleTable from}

\item{...}{Other arguments to pass to dplyr::select}
}
\value{
A qseaSet object with the sampleTable changed by a call to dplyr::select.
The sample_name and group columns will always be returned.
}
\description{
This function extends the dplyr function select to act on qseaSet sampleTable. Can also be used to rename columns.
}
\examples{
# select only a few columns, note that sample_name and group are always kept
# note here we use the `dplyr::` prefix, to ensure that `AnnotationDbi::select` isn't called. 
exampleTumourNormal \%>\% dplyr::select(type, tumour) \%>\% getSampleTable()
# select and rename a column
exampleTumourNormal \%>\% dplyr::select(age, tumour_new = tumour) \%>\% getSampleTable()
# you can use tidyselect helper functions
exampleTumourNormal \%>\% dplyr::select(matches("s")) \%>\% getSampleTable()
# negative selection also works, but not for `sample_name` and `group`:
exampleTumourNormal \%>\% dplyr::select(-matches("s")) \%>\% getSampleTable()
}
