% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{select.qseaSet}
\alias{select.qseaSet}
\title{This function extends the dplyr function select to act on qseaSet sampleTable. Can also be used to rename columns.}
\usage{
\method{select}{qseaSet}(.data, ...)
}
\arguments{
\item{.data}{A qseaSet to select columns of the sampleTable from}

\item{...}{Other arguments to pass to dplyr::select}
}
\value{
A qseaSet object with the sampleTable changed by a call to dplyr::select
}
\description{
This function extends the dplyr function select to act on qseaSet sampleTable. Can also be used to rename columns.
}
