% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{selectQset}
\alias{selectQset}
\title{This function takes a qseaSet and selects columns from its sampleTable based on a call to dplyr::select}
\usage{
selectQset(qseaSet, ...)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{...}{Other arguments to pass to dplyr::select}
}
\value{
A qseaSet object with the sampleTable selected as specified.
}
\description{
This function takes a qseaSet and selects columns from its sampleTable based on a call to dplyr::select
}
