% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{setMesaAnnoDb}
\alias{setMesaAnnoDb}
\title{This function sets a default annoDb for the annotateWindows function to use.}
\usage{
setMesaAnnoDb(annoDb)
}
\arguments{
\item{annoDb}{An annoDb package (as a string, or the object itself)}
}
\value{
None
}
\description{
For instance "org.Hs.eg.db" for human, or "org.Mm.eg.db" for mouse.
}
\seealso{
The ChIPseeker function  used by annotateWindows is annotatePeak, \link[ChIPseeker]{annotatePeak}
}
