% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{setMesaGenome}
\alias{setMesaGenome}
\title{This function sets a default genome for the annotateWindows function to use}
\usage{
setMesaGenome(genome)
}
\arguments{
\item{genome}{String corresponding to a genome.}
}
\value{
Invisibly returns \code{TRUE} on successful completion.
}
\description{
Currently supporting "hg38" and "GRCh38", this sets the TxDb and annoDb options to appropriate files.
}
\examples{
# Set the default genome to hg38
setMesaGenome("hg38")
getOption("mesa_genome")

# Use the GRCh38 alias
setMesaGenome("GRCh38")
getOption("mesa_genome")

# Reset/remove the default genome
setMesaGenome(NULL)
is.null(getOption("mesa_genome"))
}
