% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{setMesaGenome}
\alias{setMesaGenome}
\title{This function sets a default genome for the annotateWindows function to use}
\usage{
setMesaGenome(genome)
}
\arguments{
\item{genome}{String corresponding to a genome.}
}
\value{
None
}
\description{
Currently supporting "hg38" and "GRCh38", this sets the TxDb and annoDb options to appropriate files.
}
