% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{setMesaParallel}
\alias{setMesaParallel}
\title{Toggle whether to use parallelisation or not inside various functions in mesa}
\usage{
setMesaParallel(nCores = NULL, useParallel = FALSE, verbose = TRUE)
}
\arguments{
\item{nCores}{How many cores to set. Assumes you want to use MulticoreParam, to use another architecture use the useParallel argument and specify manually.}

\item{useParallel}{Boolean denoting whether or not to use parallelisation for various functions in mesa}

\item{verbose}{Boolean to determine whether to print messages or not}
}
\value{
None
}
\description{
Toggle whether to use parallelisation or not inside various functions in mesa
}
