% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrVerbs.R
\name{sort.qseaSet}
\alias{sort.qseaSet}
\title{This function extends the function sort to act on qseaSet, to reorder the names. It uses gtools::mixedsort to sort numbers correctly (i.e. 10 does not come before 2).}
\usage{
\method{sort}{qseaSet}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{A qseaSet to reorder the samples of.}

\item{decreasing}{Whether to order the qseaSet in reverse alphabetical order}

\item{...}{Other arguments to pass to gtools::mixedsort}
}
\value{
A qseaSet with the sample names reordered alphabetically
}
\description{
This function extends the function sort to act on qseaSet, to reorder the names. It uses gtools::mixedsort to sort numbers correctly (i.e. 10 does not come before 2).
}
\examples{
sort(exampleTumourNormal, decreasing = TRUE)
}
