% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{subsetWindowsBySignal}
\alias{subsetWindowsBySignal}
\title{Filter windows from a qseaSet based on expression in a subset of samples}
\usage{
subsetWindowsBySignal(
  qseaSet,
  fn,
  threshold,
  keepTrue,
  samples = NULL,
  normMethod = "nrpm",
  useGroups = FALSE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{fn}{A function that takes a row of data and returns a single value.}

\item{threshold}{The value to threshold the data rows on}

\item{keepTrue}{Whether to keep the windows passing the function, or to filter these out}

\item{samples}{A set of sample names to filter on, or a string to match in the sample names}

\item{normMethod}{The type of normalisation method to use (e.g. nrpm, beta, counts)}

\item{useGroups}{Whether to use the group argument of the sample_name to collapse replicates.}
}
\value{
A qseaSet object, with only a subset of the windows.
}
\description{
This function is designed to take a qseaSet and filter the windows by applying a function to each window, followed by thresholding either windows above or below that cutoff.
For instance, we can keep only regions that have a median normalised reads per million value above 1, or those where the minimum beta value is below 0.5.
Which samples to apply the filter based on can be specified with the \code{samples} argument, either as a list of samples or a string.
}
