% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{subsetWindowsBySignal}
\alias{subsetWindowsBySignal}
\title{Filter windows from a qseaSet based on expression in a subset of samples}
\usage{
subsetWindowsBySignal(
  qseaSet,
  fn,
  threshold,
  aboveThreshold,
  samples = NULL,
  normMethod = "nrpm",
  useGroupMeans = FALSE
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{fn}{A function that takes a row of data and returns a single value.}

\item{threshold}{The value to threshold the \code{fn} output value on.}

\item{aboveThreshold}{A logical value indicating whether to keep the windows with \code{fn} output value above or equal to \code{threshold} (TRUE), or below \code{threshold} (FALSE).}

\item{samples}{A set of sample names to filter on, or a string to match in the sample names.}

\item{normMethod}{The type of normalisation method to use (e.g. nrpm, beta, counts).}

\item{useGroupMeans}{Whether to use the group argument of the sample table to collapse replicates.}
}
\value{
A qseaSet object, with only a subset of the windows.
}
\description{
This function is designed to take a qseaSet and filter the windows by applying a function to each window, followed by thresholding either windows above or below that cutoff.
For instance, we can keep only regions that have a median normalised reads per million value above 1, or those where the minimum beta value is below 0.5.
Which samples to apply the filter based on can be specified with the \code{samples} argument, either as a list of samples or a string.
}
\examples{
# keep only windows with a mean nrpm value above 1 across all samples
exampleTumourNormal \%>\% subsetWindowsBySignal(fn = mean, threshold = 1, aboveThreshold = TRUE, normMethod = "nrpm")
# keep only windows with a median beta value below 0.25 in the normal samples
exampleTumourNormal \%>\% subsetWindowsBySignal(fn = median, threshold = 0.25, aboveThreshold = FALSE, normMethod = "beta", samples = "_N")

}
