% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{subsetWindowsOverBackground}
\alias{subsetWindowsOverBackground}
\title{This function takes a qseaSet and finds which windows have more reads than would be expected if the reads followed a Poisson distribution over the whole genome.}
\usage{
subsetWindowsOverBackground(
  qseaSet,
  keepAbove = FALSE,
  samples = NULL,
  numWindows = NULL,
  recalculateNumWindows = TRUE,
  fdrThres = 0.01,
  numAbove = 1
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{keepAbove}{Boolean, if true, then keep only the windows that are above the background, if false then remove them.}

\item{samples}{A vector of samples to remove, or a character string to match and remove.}

\item{numWindows}{The number of windows in the whole genome.}

\item{recalculateNumWindows}{Whether to recalculate the number of windows based on the whole genome.}

\item{fdrThres}{What FDR p-value to filter windows out based on.}

\item{numAbove}{How many samples need to be above the background level in a window to keep/remove it.}
}
\value{
A qseaSet object, with only a subset of the windows.
}
\description{
This function takes a qseaSet and finds which windows have more reads than would be expected if the reads followed a Poisson distribution over the whole genome.
}
