% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{summariseAcrossWindows}
\alias{summariseAcrossWindows}
\title{This function takes a qseaSet and calculates a summary statistic across the windows.}
\usage{
summariseAcrossWindows(
  qseaSet,
  windowsToUse = NULL,
  fn = mean,
  addSampleTable = TRUE,
  normMethod = c("nrpm", "beta"),
  naMethod = "na.rm",
  minEnrichment = 3,
  suffix = "",
  fnName = NULL
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{windowsToUse}{A GRanges object (or dataframe coercible to one) containing the windows to summarise over. If missing, will use all the regions in the qseaSet.}

\item{fn}{A function to apply across the windows. e.g. mean, median, sd.}

\item{addSampleTable}{A boolean with whether to add the sampleTable on to the output table}

\item{normMethod}{One or more normalisation methods to use, e.g. "nrpm" or "beta" or c("nrpm", "beta").}

\item{naMethod}{What method to use to deal with NA values (for beta values). Options are "drop" or "na.rm" (default).}

\item{minEnrichment}{For beta values only, the minimum number of reads required for a window to be fully methylated, qsea replaces with NA below this value.}

\item{suffix}{A suffix for adding to the new columns of data, to clarify what regions for instance.}

\item{fnName}{Name of the function. Should only be necessary to use if you are doing something unusual, otherwise detected automatically.}
}
\value{
A table of data, one row per sample, with columns indicating the application of the summary statistic to the
}
\description{
This function takes a qseaSet and calculates a summary statistic across the windows.
}
