% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{summariseAcrossWindows}
\alias{summariseAcrossWindows}
\title{This function takes a qseaSet and calculates a summary statistic across the windows.}
\usage{
summariseAcrossWindows(
  qseaSet,
  regionsToOverlap = NULL,
  fn = mean,
  addSampleTable = TRUE,
  normMethod = c("nrpm", "beta"),
  naMethod = "na.rm",
  minEnrichment = 3,
  suffix = "",
  fnName = NULL
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{regionsToOverlap}{A GRanges object (or coercible dataframe) containing the windows to summarise over. If missing, will use all the regions in the qseaSet.}

\item{fn}{A function to apply across the windows. e.g. mean, median, sd.}

\item{addSampleTable}{A boolean with whether to add the sampleTable on to the output table}

\item{normMethod}{One or more normalisation methods to use, e.g. "nrpm" or "beta" or c("nrpm", "beta").}

\item{naMethod}{What method to use to deal with NA values (for beta values). Options are "drop" or "na.rm" (default).}

\item{minEnrichment}{For beta values only, the minimum number of reads required for a window to be fully methylated, qsea replaces with NA below this value.}

\item{suffix}{A suffix for adding to the new columns of data, to clarify what regions for instance.}

\item{fnName}{Name of the function. Should only be necessary to use if you are doing something unusual, otherwise detected automatically.}
}
\value{
A table of data, one row per sample, with columns indicating the number of windows considered
(which may be less for beta values due to application of NA values), the summary statistic and optionally the sampleTable columns.
}
\description{
This function takes a qseaSet and calculates a summary statistic across the windows.
}
\examples{
# median NRPM and beta values across all windows
exampleTumourNormal \%>\% summariseAcrossWindows(fn = mean)
# maximum NRPM across a smaller region of the genome
rgns <- data.frame(seqnames = 7, start = 25002001, end = 25017900
exampleTumourNormal \%>\% summariseAcrossWindows(regionsToOverlap = rgns, fn = max, normMethod = "nrpm")
}
