% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseDMRs.R
\name{summariseDMRsByContrast}
\alias{summariseDMRsByContrast}
\title{Summarise DMRs by contrast}
\usage{
summariseDMRsByContrast(
  DMRtable,
  FDRthres = 0.05,
  log2FCthres = 0,
  deltaBetaThres = 0
)
}
\arguments{
\item{DMRtable}{\code{data.frame} or \code{GRanges} of DMR results, typically from \code{\link[=calculateDMRs]{calculateDMRs()}}.
If in wide format, it will be converted with \code{\link[=pivotDMRsLonger]{pivotDMRsLonger()}}.}

\item{FDRthres}{\code{numeric(1)} False discovery rate threshold.}

\item{log2FCthres}{\code{numeric(1)} Absolute log2 fold-change threshold.}

\item{deltaBetaThres}{\code{numeric(1)} Absolute delta-beta threshold.}
}
\value{
A tibble with one row per contrast and columns \code{nUp} and \code{nDown}.
}
\description{
Count the number of up- and down-regulated windows per contrast.
Internally transforms the DMR results to long format.
}
\examples{
data(exampleTumourNormal, package = "mesa")
exampleTumourNormal \%>\%
   calculateDMRs(variable = "type", contrasts = "LUAD_vs_NormalLung", FDRthres = 0.1) \%>\%
   summariseDMRsByContrast(FDRthres = 0.1)

}
\seealso{
\code{\link[=pivotDMRsLonger]{pivotDMRsLonger()}}, \code{\link[=summariseDMRsByGene]{summariseDMRsByGene()}}

Other DMR-helpers: 
\code{\link{pivotDMRsLonger}()},
\code{\link{summariseDMRsByGene}()},
\code{\link{writeDMRsToBed}()},
\code{\link{writeDMRsToExcel}()}
}
\concept{DMR-helpers}
