% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseDMRs.R
\name{summariseDMRsByGene}
\alias{summariseDMRsByGene}
\title{This function summarises a data frame by gene (using annotateData)}
\usage{
summariseDMRsByGene(DMRtable)
}
\arguments{
\item{DMRtable}{A data frame or GRanges object, may already be annotated}
}
\description{
This function summarises a data frame by gene (using annotateData)
}
