% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qseaExtra.R
\name{writeBigWigs}
\alias{writeBigWigs}
\title{This function takes a qseaSet and writes individual bigWig files with the coverage over each window.}
\usage{
writeBigWigs(
  qseaSet,
  folderName,
  normMethod = "nrpm",
  useGroupMeans = FALSE,
  naVal = -1
)
}
\arguments{
\item{qseaSet}{The qseaSet object.}

\item{folderName}{A folder to save the output bigWig files into}

\item{normMethod}{What normalisation method to use (e.g. nrpm or beta)}

\item{useGroupMeans}{Whether to average over the replicates in the group}

\item{naVal}{A value to replace NA values with (for beta values).}
}
\value{
A set of bigWig files for each sample in the qseaSet, with coverage over all the windows in the qseaSet
}
\description{
This function takes a qseaSet and writes individual bigWig files with the coverage over each window.
}
