% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseDMRs.R
\name{writeDMRsToBed}
\alias{writeDMRsToBed}
\title{Write one or more DMRs to a set of bed files}
\usage{
writeDMRsToBed(dataTable, folder, fdrThres = 0.05)
}
\arguments{
\item{dataTable}{An (annotated) data table with contrasts}

\item{folder}{The folder to place the bed files.}

\item{fdrThres}{FDR rate threshold to apply to each contrast}
}
\value{
Returns the original table invisibly, so can be used in a pipe.
}
\description{
This writes a set of bed files with the results, with one bed file per contrast.
}
\examples{
exampleTumourNormal \%>\%
  calculateDMRs(variable = "tumour", contrasts = "all") \%>\%
  annotateWindows(TxDb = "TxDb.Hsapiens.UCSC.hg38.knownGene", annoDb = "org.Hs.eg.db") \%>\%
  writeDMRsToBed(tempdir())
}
